### 2020-07-28 #2529 @whedon
[PRE REVIEW]: matador: a Python library for analysing, curating and performing high-throughput density-functional theory calculations
**Submitting author:** @ml-evs (<a href="http://orcid.org/0000-0002-1182-9098">Matthew L Evans</a>)
**Repository:** <a href="https://github.com/ml-evs/matador" target ="_blank">https://github.com/ml-evs/matador</a>
**Version:** v0.9.6
