### 2020-07-28 1d8259b
BLD: amend test location in setup.py
