### 2020-07-28 91126e1
Updates to README
- Added citation for Zenodo
- Add notes about dependencies
- Added caveats to use cases
