+++ README.rst
@@ -6,7 +6,7 @@ matador
 |GH Actions| |Coverage Status| |Documentation Status|
 
 matador is an aggregator, manipulator and runner of first-principles calculations, written with a bent towards battery electrode materials.
-The source can be found on `GitHub <https://github.com/ml-evs/matador>`_ and online documentation is hosted on `ReadTheDocs <https://docs.matador.science>`_.
+The source can be found on `GitHub <https://github.com/ml-evs/matador>`_ and online documentation is hosted at `ReadTheDocs <https://docs.matador.science>`_.
 
 Example Jupyter notebooks and tutorials can be found `online <https://docs.matador.science/en/latest/examples_index.html>`_ or in the ``examples/`` folder of the matador source code.
 
@@ -21,6 +21,8 @@ Installation
 ------------
 
 In the simplest case (e.g. you already have Python 3.6+ set up), ``pip install matador-db`` is sufficient to get up and running, preferably in a fresh virtual environment. More detailed instructions can be found in the `Installation instructions <https://docs.matador.science/en/latest/install.html>`_.
+
+Tests can be run on your local machine with ``python -m unittest discover -v -b`` or simply with ``py.test`` (required pytest to be installed); in order to test CASTEP-running functionality, the tests will look for an MPI-enabled executable named ``castep`` on your ``$PATH``.
 
 Upgrading to the latest version should be as simple as ``pip install -U matador-db``.
 
@@ -41,8 +43,10 @@ The API has many features that can be explored in the examples and API documenta
 - Scraping of CASTEP (and Quantum Espresso) input/output files into flexible Python dictionaries/models.
 - The creation and curation of MongoDB collections of geometry optimisation, calculations, with a powerful querying CLI/API.
 - Customisable, publication-ready plots for all models, e.g. phase diagrams, PDF, PXRD, voltage profiles, electronic/vibrational bandstructures etc.
-- Automated high-throughput geometry optimisations, electronic and vibrational properties using CASTEP (and Quantum Espresso) with ``run3``. Tested on several supercomputers.
+- High-throughput geometry optimisations, electronic and vibrational properties using CASTEP (and Quantum Espresso) with ``run3``. Tested on several supercomputers. ``run3`` is designed primarily for simple workflows and offers little in the way of tools for creating complex workflows; if this is your use case, then check out some of the other codes listed below.
 - Creation of phase diagrams and electrochemical voltage profiles from the results of DFT calculations.
+
+This functionality is achieved by interfacing with much of the standard scientific Python stack (`NumPy <https://numpy.org>`_, `SciPy <https://scipy.org>`_, `matplotlib <https://matplotlib.org>`_), some more materials-specific packages (`spglib <https://github.com/atztogo/spglib/>`_, `SeeK-path <https://github.com/giovannipizzi/seekpath>`_, `periodictable <https://github.com/pkienzle/periodictable>`_) and other general packages (`pymongo <https://github.com/mongodb/mongo-python-driver>`_, `python-ternary <https://github.com/marcharper/python-ternary>`_, `numba <https://numba.org`_).
 
 Similar packages
 ----------------
@@ -50,11 +54,18 @@ Similar packages
 This package is by no means unique in its functionality or goals. Below is a list of similar packages and an overview of where they overlap with ``matador``:
 
 - `ASE <https://wiki.fysik.dtu.dk/ase/>`_: manipulation of structures, parsing and exporting files, running jobs and local databases with ``ase-db``. An interface is provided to ASE's ``Atoms`` object.
[…]
