### 2020-07-28 #2529 @jgostick
@mkhorton are you offering to editor this one?  If so, go for it.
