### 2020-07-28 #2529 @mkhorton
I was offering to review in response to @ml-evs potential reviewers list. I'd be happy to be editor, but I haven't previously taken any editorial role in JOSS
