### 2020-07-28 c7c4b76
MAINT: deprecated unfinished Electrode functionality
