### 2020-07-28 dd10584
MAINT: added some headers and switched to relative imports in module init
