+++ matador/scrapers/__init__.py
@@ -21,9 +21,9 @@ __all__ = ['res2dict', 'cell2dict', 'param2dict', 'castep2dict', 'bands2dict', '
 __author__ = 'Matthew Evans'
 __maintainer__ = 'Matthew Evans'
 
-from matador.scrapers.castep_scrapers import res2dict, cell2dict, param2dict
-from matador.scrapers.castep_scrapers import castep2dict, bands2dict, arbitrary2dict
-from matador.scrapers.castep_scrapers import phonon2dict, phonon_dos2dict, optados2dict, usp2dict
-from matador.scrapers.qe_scrapers import pwout2dict
-from matador.scrapers.magres_scrapers import magres2dict
-from matador.scrapers.cif_scraper import cif2dict
+from .castep_scrapers import res2dict, cell2dict, param2dict
+from .castep_scrapers import castep2dict, bands2dict, arbitrary2dict
+from .castep_scrapers import phonon2dict, phonon_dos2dict, optados2dict, usp2dict
+from .qe_scrapers import pwout2dict
+from .magres_scrapers import magres2dict
+from .cif_scraper import cif2dict
