+++ matador/orm/__init__.py
@@ -1,0 +1,9 @@
+# coding: utf-8
+# Distributed under the terms of the MIT License.
+
+__all__ = ["DataContainer"]
+
+__author__ = 'Matthew Evans'
+__maintainer__ = 'Matthew Evans'
+
+from .orm import DataContainer
