### 2020-07-28 2fbb8ed
DOCS: fixed some docstrings
