### 2020-07-28 11a1ccb
v0.9.6 (#87)
* Improvements to ASE and pymatgen interoperability (#80)
* Fixed bug in matador.hull.TemperatureDependentHull which would crash when not provided a list of temperatures (#82).
* Added plotting functions for magres data, and improved its handling inside matador.crystal.Crystal (#79).
* Added code of conduct
