### 2020-07-29 1edf95d
MAINT: improve error messages and removed some stub code
