### 2020-07-29 93957fa
FIX: incorrectly displayed error due to bad finally block
