### 2020-07-29 ad017aa
Two fixes that were causing misleading logs:
- at somepoint castep_singleshot stopped returning its status, so workflows were reported as failing
- sources printed from Crystal objects were incorrect as they were overwritten by parameter files
