### 2020-07-29 cf4271c
Some tidying and more minor fixes (#88)
* FIX: broken import

* MAINT: tidying old code

* FIX: title for example notebook

* BLD: extra deps required when building docs

* MAINT: improve error messages and removed some stub code

* FIX: incorrectly displayed error due to bad finally block

* MAINT: removed old flag from cell conversion scripts

* FIX: Two fixes that were causing misleading logs:
    - at somepoint castep_singleshot stopped returning its status, so workflows were reported as failing
    - sources printed from Crystal objects were incorrect as they were overwritten by parameter
[…]
