+++ readthedocs.yml
@@ -7,6 +7,7 @@ python:
     - viz
     - db
     - crystal
+    - optional
 
 build:
     image: latest
