### 2020-07-29 85e12da
v0.9.7
-Bug fixes to problems introduced in 0.9.6 (#88)
    - Cosmetic fixes to logging and misleading status reports in workflows
    - Improved some error messages
