### 2020-07-30 34a3162
ENH: scrape units from magres files and rename asymm/aniso to shift from shielding
