+++ tests/test_file_io.py
@@ -1504,6 +1504,11 @@ class ScraperMiscTest(MatadorUnitTest):
             ),
         )
 
+        self.assertEqual(magres_dict["magres_units"]["ms"], "ppm")
+        self.assertEqual(magres_dict["magres_units"]["lattice"], "Angstrom")
+        self.assertEqual(magres_dict["magres_units"]["atom"], "Angstrom")
+        self.assertEqual(magres_dict["magres_units"]["sus"], "10^-6.cm^3.mol^-1")
+
         np.testing.assert_almost_equal(
             magres_dict["susceptibility_tensor"],
             [
@@ -1563,7 +1568,7 @@ class ScraperMiscTest(MatadorUnitTest):
             decimal=1,
         )
         np.testing.assert_almost_equal(
-            magres_crystal["chemical_shielding_anisos"],
+            magres_crystal["chemical_shift_anisoss"],
             [
                 9.4,
                 4.4,
@@ -1589,7 +1594,7 @@ class ScraperMiscTest(MatadorUnitTest):
             decimal=1,
         )
         np.testing.assert_almost_equal(
-            magres_crystal["chemical_shielding_asymmetries"],
+            magres_crystal["chemical_shift_asymmmetries"],
             [
                 0.33,
                 0.76,
@@ -1621,16 +1626,21 @@ class ScraperMiscTest(MatadorUnitTest):
                 magres_crystal["chemical_shielding_isos"][ind],
             )
             self.assertEqual(
-                atom["chemical_shielding_aniso"],
-                magres_crystal["chemical_shielding_anisos"][ind],
+                atom["chemical_shift_anisos"],
+                magres_crystal["chemical_shift_anisoss"][ind],
             )
             self.assertEqual(
-                atom["chemical_shielding_asymmetry"],
-                magres_crystal["chemical_shielding_asymmetries"][ind],
+                atom["chemical_shift_asymmmetry"],
+                magres_crystal["chemical_shift_asymmmetries"][ind],
             )
 
         self.assertEqual(magres_crystal["calculator"], "CASTEP")
         self.assertEqual(magres_crystal["calculator_version"], "18.1")
+
+        self.assertEqual(magres_crystal["magres_units"]["ms"], "ppm")
+        self.assertEqual(magres_crystal["magres_units"]["lattice"], "Angstrom")
+        self.assertEqual(magres_crystal["magres_units"]["atom"], "Angstrom")
+        self.assertEqual(magres_crystal["magres_units"]["sus"], "10^-6.cm^3.mol^-1")
 
     def test_pwscfout(self):
         pwout_fname = REAL_PATH + "data/NaP.out"
