### 2020-07-30 08ae2da
ENH: use units in magres plot and update example
