### 2020-07-30 7906899
ENH: use units in magres plot and update example
