### 2020-07-30 23003ae
ENH: use units in magres plot and update example
