### 2020-07-30 fb2204f
TST: test example notebooks in CI
