+++ .github/workflows/ci.yml
@@ -55,7 +55,14 @@ jobs:
     - name: Test with pytest
       run: |
         py.test -rs --cov-config .coveragerc --cov=./matador/ --cov-report=xml
+    - name: Test notebook examples
+      run: |
+        py.test -rs -vvv --nbval --current-env --cov-config .coveragerc --cov=./matador/ --cov-append --cov-report=xml examples/magres_plotting/
+        py.test -rs -vvv --nbval-lax --current-env --cov-config .coveragerc --cov=./matador/ --cov-append --cov-report=xml examples/pymatgen_and_ase_interfaces/
+        py.test -rs -vvv --nbval-plot_styleslax --current-env --cov-config .coveragerc --cov=./matador/ --cov-append --cov-report=xml examples/plot_styles/
+        py.test -rs -vvv --nbval --current-env --cov-config .coveragerc --cov=./matador/ --cov-append --cov-report=xml examples/voltage/
     - name: Upload coverage to Codecov
+      if: matrix.python-version == 3.8 && github.repository == 'ml-evs/matador'
       uses: codecov/codecov-action@v1
       with:
         token: ${{ secrets.CODECOV_TOKEN }}
