### 2020-07-30 1de6b01
Improvements to magres functionality (#90)
* ENH: scrape units from magres files and rename asymm/aniso to shift from shielding
* ENH: use units in magres plot and update example
