### 2020-07-30 a22aef4
BLD: update matplotlib configs to avoid warnings in nbval CI
