+++ .github/workflows/ci.yml
@@ -51,7 +51,7 @@ jobs:
       run: |
         cp tests/data/matador_pipelines_conf.yml $HOME/.matadorrc
         python -m pip install --upgrade pip
-        pip install -e .[test,plotting,crystal,optional]
+        pip install -e .[test,crystal,optional]
     - name: Test with pytest
       run: |
         py.test -rs --cov-config .coveragerc --cov=./matador/ --cov-report=xml
@@ -59,7 +59,7 @@ jobs:
       run: |
         py.test -rs -vvv --nbval --current-env --cov-config .coveragerc --cov=./matador/ --cov-append --cov-report=xml examples/magres_plotting/
         py.test -rs -vvv --nbval-lax --current-env --cov-config .coveragerc --cov=./matador/ --cov-append --cov-report=xml examples/pymatgen_and_ase_interfaces/
-        py.test -rs -vvv --nbval-plot_styleslax --current-env --cov-config .coveragerc --cov=./matador/ --cov-append --cov-report=xml examples/plot_styles/
+        py.test -rs -vvv --nbval-lax --current-env --cov-config .coveragerc --cov=./matador/ --cov-append --cov-report=xml examples/plot_styles/
         py.test -rs -vvv --nbval --current-env --cov-config .coveragerc --cov=./matador/ --cov-append --cov-report=xml examples/voltage/
     - name: Upload coverage to Codecov
       if: matrix.python-version == 3.8 && github.repository == 'ml-evs/matador'
