### 2020-07-30 59f6414
Only test notebooks with py38
