### 2020-07-30 f72edf3
Run example notebooks in CI (#91)
* DOCS: various examples updates
* TST: test example notebooks in CI
* TST: bump requirements
* BLD: update matplotlib configs to avoid warnings in nbval CI
* BLD: Only test notebooks with py38
