+++ .github/workflows/ci.yml
@@ -51,11 +51,19 @@ jobs:
       run: |
         cp tests/data/matador_pipelines_conf.yml $HOME/.matadorrc
         python -m pip install --upgrade pip
-        pip install -e .[test,plotting,crystal,optional]
+        pip install -e .[test,crystal,optional]
     - name: Test with pytest
       run: |
         py.test -rs --cov-config .coveragerc --cov=./matador/ --cov-report=xml
+    - name: Test notebook examples
+      if: matrix.python-version == 3.8 && github.repository == 'ml-evs/matador'
+      run: |
+        py.test -rs -vvv --nbval --current-env --cov-config .coveragerc --cov=./matador/ --cov-append --cov-report=xml examples/magres_plotting/
+        py.test -rs -vvv --nbval --current-env --cov-config .coveragerc --cov=./matador/ --cov-append --cov-report=xml examples/voltage/
+        py.test -rs -vvv --nbval-lax --current-env --cov-config .coveragerc --cov=./matador/ --cov-append --cov-report=xml examples/pymatgen_and_ase_interfaces/
+        py.test -rs -vvv --nbval-lax --current-env --cov-config .coveragerc --cov=./matador/ --cov-append --cov-report=xml examples/plot_styles/
     - name: Upload coverage to Codecov
+      if: matrix.python-version == 3.8 && github.repository == 'ml-evs/matador'
       uses: codecov/codecov-action@v1
       with:
         token: ${{ secrets.CODECOV_TOKEN }}
