### 2020-08-06 ac2eec8
ENH: add numerical tolerance to gap finding
