### 2020-08-06 fa035b7
FIX: plot_window arg being properly parsed when provided as single value
