### 2020-08-06 cad38ec
MAINT: disable band_reorder with pdis as it gives garbage results
