### 2020-08-06 6d5eea7
ENH: scrape old 'bs_*' keywords as their new 'spectral_*' counterparts (closes #93)
