+++ matador/scrapers/castep_scrapers.py
@@ -246,6 +246,8 @@ def cell2dict(seed, db=False, lattice=True, positions=True, **kwargs):
         elif 'kpoints_mp_spacing' in line.lower() or 'kpoint_mp_spacing' in line.lower():
             if 'spectral_kpoints_mp_spacing' in line.lower() or 'spectral_kpoint_mp_spacing' in line.lower():
                 cell['spectral_kpoints_mp_spacing'] = f90_float_parse(line.split()[-1])
+            elif 'bs_kpoints_mp_spacing' in line.lower() or 'bs_kpoint_mp_spacing' in line.lower():
+                cell['spectral_kpoints_mp_spacing'] = f90_float_parse(line.split()[-1])
             elif 'supercell_kpoints_mp_spacing' in line.lower() or 'supercell_kpoint_mp_spacing' in line.lower():
                 cell['supercell_kpoints_mp_spacing'] = f90_float_parse(line.split()[-1])
             elif 'phonon_kpoints_mp_spacing' in line.lower() or 'phonon_kpoint_mp_spacing' in line.lower():
@@ -256,6 +258,8 @@ def cell2dict(seed, db=False, lattice=True, positions=True, **kwargs):
                 cell['kpoints_mp_spacing'] = f90_float_parse(line.split()[-1])
         elif 'kpoints_mp_grid' in line.lower() or 'kpoint_mp_grid' in line.lower():
             if 'spectral_kpoints_mp_grid' in line.lower() or 'spectral_kpoint_mp_grid' in line.lower():
+                cell['spectral_kpoints_mp_grid'] = list(map(int, line.split()[-3:]))
+            if 'bs_kpoints_mp_grid' in line.lower() or 'bs_kpoint_mp_grid' in line.lower():
                 cell['spectral_kpoints_mp_grid'] = list(map(int, line.split()[-3:]))
             # these two keywords do not have a corresponding "kpoints" alias (instead of kpoint)
             # so must remain unpluralised (see below for other phonon keyword exceptions)
@@ -268,6 +272,8 @@ def cell2dict(seed, db=False, lattice=True, positions=True, **kwargs):
         elif 'kpoints_mp_offset' in line.lower() or 'kpoint_mp_offset' in line.lower():
             if 'spectral_kpoints_mp_offset' in line.lower() or 'spectral_kpoint_mp_offset' in line.lower():
                 cell['spectral_kpoints_mp_offset'] = list(map(f90_float_parse, line.split()[-3:]))
+            if 'bs_kpoints_mp_offset' in line.lower() or 'bs_kpoint_mp_offset' in line.lower():
+                cell['spectral_kpoints_mp_offset'] = list(map(f90_float_parse, line.split()[-3:]))
             elif 'phonon_kpoints_mp_offset' in line.lower() or 'phonon_kpoint_mp_offset' in line.lower():
                 # this is a special case where phonon_kpointS_mp_offset doesn't exist
                 cell['phonon_kpoint_mp_offset'] = list(map(f90_float_parse, line.split()[-3:]))
@@ -276,12 +282,12 @@ def cell2dict(seed, db=False, lattice=True, positions=True, **kwargs):
             else:
                 cell['kpoints_mp_offset'] = list(map(f90_float_parse, line.split()[-3:]))
         elif '%block' in line.lower() and ('_kpoints_path' in line.lower() or '_kpoint_path' in line.lower()):
-            if 'spectral' in line.lower():
+            if 'spectral_' in line.lower() or "bs_" in line.lower():
                 key = 'spectral_kpoints_path'
-            elif 'phonon_fine' in line.lower():
+            elif 'phonon_fine_' in line.lower():
                 key = 'phonon_fine_kpoint_path'
             else:
-                raise RuntimeError('Found unknown kpoint path block.')
[…]
