### 2020-08-06 49838f0
MAINT: if only a single element is found in  calc, use mpl color cycle rather than element colors
