### 2020-08-06 a4f5a0e
MAINT: handle empty lines in species_pot block
