### 2020-08-06 9f50aec
FIX: catch errors in cell parsing, rather than passing them along
