### 2020-08-06 1b71d5f
ENH: added --projector_colours to dispersion for colour override functionality
