### 2020-08-06 f3c4d62
MAINT: made dispersion_plot(...) and dos_plot(...) functions more API-friendly
