+++ matador/plotting/spectral_plotting.py
@@ -23,6 +23,10 @@ from matador.utils.chem_utils import INVERSE_CM_TO_EV
 
 __all__ = ['plot_spectral']
 
+PROJECTOR_MIN = 1e-3
+SPIN_UP_COLOUR = "firebrick"
+SPIN_DOWN_COLOUR = "dodgerblue"
+
 
 @plotting_function
 def plot_spectral(seeds, **kwargs):
@@ -166,14 +170,19 @@ def plot_spectral(seeds, **kwargs):
             kwargs['plot_pdos'] = any([any([os.path.isfile('{}.{}'.format(seed, ext)) for ext in exts]) for seed in seeds])
 
     figsize = kwargs['figsize']
+    # need to rasterize pdis plots, which we'll force to have negative zorders
+    rasterized = kwargs.get('plot_pdis', False)
+    subplot_kw = {"rasterized": rasterized}
+    if rasterized:
+        subplot_kw["rasterization_zorder"] = -1e3
     if kwargs['plot_bandstructure'] and not kwargs['plot_dos']:
         if figsize is None:
             figsize = (7, 6)
-        fig, ax_dispersion = plt.subplots(figsize=figsize)
+        fig, ax_dispersion = plt.subplots(figsize=figsize, subplot_kw=subplot_kw)
     elif kwargs['plot_bandstructure'] and kwargs['plot_dos']:
         if figsize is None:
             figsize = (10, 6)
-        fig, ax_grid = plt.subplots(1, 3, figsize=figsize, sharey=True,
+        fig, ax_grid = plt.subplots(1, 3, figsize=figsize, sharey=True, subplot_kw=subplot_kw,
                                     gridspec_kw={'width_ratios': [4, 2, 1],
                                                  'wspace': 0.1,
                                                  'left': 0.15})
@@ -183,7 +192,7 @@ def plot_spectral(seeds, **kwargs):
     elif not kwargs['plot_bandstructure'] and kwargs['plot_dos']:
         if figsize is None:
             figsize = (9, 4)
-        fig, ax_dos = plt.subplots(1, figsize=figsize)
+        fig, ax_dos = plt.subplots(1, figsize=figsize, subplot_kw=subplot_kw)
     else:
         raise RuntimeError("No plots requested, please set either plot_dos or plot_bandstructure to True!")
 
@@ -243,12 +252,13 @@ def plot_spectral(seeds, **kwargs):
             plt.show()
 
 
-def dispersion_plot(seeds, ax_dispersion, kwargs, bbox_extra_artists):
+@plotting_function
+def dispersion_plot(seeds, ax_dispersion, kwargs, bbox_extra_artists=None):
     """ Plot a dispersion/bandstructure on the given axis. Will detect
     and projected dispersion data automatically.
 
     Parameters:
-        seeds (list): the seednames of the data to plot.
+        seeds (str or list): the seedname(s) of the data to plot.
         ax_dispersion (matplotlib.Axes): the axis to plot on.
         kwargs (dict): any plotting keywords (from e.g. dispersion script).
         bbox_extra_artists (list): a list to which to append legends.
@@ -260,6 +270,13 @@ def dispersion_plot(seeds, ax_dispersion, kwargs, bbox_extra_artists):
     from cycler import cycler
     import matplotlib.pyplot as plt
     plotted_pdis = False
+
+    if not isinstance(seeds, list):
+        seeds = [seeds]
+
+    if bbox_extra_artists is None:
+        bbox_extra_artists = []
+
     for seed_ind, seed in enumerate(seeds):
 
         if isinstance(seed, dict):
@@ -286,7 +303,7 @@ def dispersion_plot(seeds, ax_dispersion, kwargs, bbox_extra_artists):
[…]
