### 2020-08-06 c2a11a4
MAINT: improvements to PDIS rendering
