### 2020-08-06 db0bee0
MAINT: force enable rasterisation for PDIS
