### 2020-08-06 fc470ec
Spectral plotting improvements (#94)
* ENH: add numerical tolerance to gap finding

* MAINT: removed defunct matplotlib parameter

* FIX: plot_window arg being properly parsed when provided as single value

* MAINT: disable band_reorder with pdis as it gives garbage results

* ENH: scrape old 'bs_*' keywords as their new 'spectral_*' counterparts (closes #93)

* MAINT: if only a single element is found in  calc, use mpl color cycle rather than element colors

* MAINT: handle empty lines in species_pot block

* FIX: catch errors in cell parsing, rather than passing them along

* DOCS: added sub
[…]
