+++ matador/scrapers/castep_scrapers.py
@@ -187,6 +187,10 @@ def cell2dict(seed, db=False, lattice=True, positions=True, **kwargs):
             cell['species_pot'] = dict()
             i = 1
             while 'endblock' not in flines[line_no + i].lower():
+                # handle blank lines in species pot
+                if not flines[line_no + i].split():
+                    i += 1
+                    continue
                 if db:
                     species = flines[line_no+i].split()[0]
                     pspot_string = flines[line_no+i].split()[1].split('/')[-1]
@@ -199,6 +203,8 @@ def cell2dict(seed, db=False, lattice=True, positions=True, **kwargs):
                     else:
                         cell['species_pot'][flines[line_no + i].split()[0]] = flines[line_no + i].split()[1]
                 i += 1
+            if not cell['species_pot']:
+                cell.pop('species_pot')
         elif '%block cell_constraints' in line.lower():
             cell['cell_constraints'] = []
             for j in range(2):
@@ -246,6 +252,8 @@ def cell2dict(seed, db=False, lattice=True, positions=True, **kwargs):
         elif 'kpoints_mp_spacing' in line.lower() or 'kpoint_mp_spacing' in line.lower():
             if 'spectral_kpoints_mp_spacing' in line.lower() or 'spectral_kpoint_mp_spacing' in line.lower():
                 cell['spectral_kpoints_mp_spacing'] = f90_float_parse(line.split()[-1])
+            elif 'bs_kpoints_mp_spacing' in line.lower() or 'bs_kpoint_mp_spacing' in line.lower():
+                cell['spectral_kpoints_mp_spacing'] = f90_float_parse(line.split()[-1])
             elif 'supercell_kpoints_mp_spacing' in line.lower() or 'supercell_kpoint_mp_spacing' in line.lower():
                 cell['supercell_kpoints_mp_spacing'] = f90_float_parse(line.split()[-1])
             elif 'phonon_kpoints_mp_spacing' in line.lower() or 'phonon_kpoint_mp_spacing' in line.lower():
@@ -256,6 +264,8 @@ def cell2dict(seed, db=False, lattice=True, positions=True, **kwargs):
                 cell['kpoints_mp_spacing'] = f90_float_parse(line.split()[-1])
         elif 'kpoints_mp_grid' in line.lower() or 'kpoint_mp_grid' in line.lower():
             if 'spectral_kpoints_mp_grid' in line.lower() or 'spectral_kpoint_mp_grid' in line.lower():
+                cell['spectral_kpoints_mp_grid'] = list(map(int, line.split()[-3:]))
+            if 'bs_kpoints_mp_grid' in line.lower() or 'bs_kpoint_mp_grid' in line.lower():
                 cell['spectral_kpoints_mp_grid'] = list(map(int, line.split()[-3:]))
             # these two keywords do not have a corresponding "kpoints" alias (instead of kpoint)
             # so must remain unpluralised (see below for other phonon keyword exceptions)
@@ -268,6 +278,8 @@ def cell2dict(seed, db=False, lattice=True, positions=True, **kwargs):
         elif 'kpoints_mp_offset' in line.lower() or 'kpoint_mp_offset' in line.lower():
             if 'spectral_kpoints_mp_offset' in line.lower() or 'spectral_kpoint_mp_offset' in line.lower():
                 cell['spectral_kpoints_mp_offset'] = list(map(f90_float_parse, line.split()[-3:]))
+            if 'bs_kpoints_mp_offset' in line.lower() or 'bs_kpoint_mp_offset' in line.lower():
+                cell['spectral_kpoints_mp_offset'] = list(map(f90_float_parse, line.split()[-3:]))
             elif 'phonon_kpoints_mp_offset' in line.lower() or 'phonon_kpoint_mp_offset' in line.lower():
[…]
