+++ matador/plotting/spectral_plotting.py
@@ -23,6 +23,10 @@ from matador.utils.chem_utils import INVERSE_CM_TO_EV
 
 __all__ = ['plot_spectral']
 
+PROJECTOR_MIN = 1e-5
+SPIN_UP_COLOUR = "firebrick"
+SPIN_DOWN_COLOUR = "dodgerblue"
+
 
 @plotting_function
 def plot_spectral(seeds, **kwargs):
@@ -68,7 +72,6 @@ def plot_spectral(seeds, **kwargs):
         filename (str): filename for figure saving.
         cmap (str): matplotlib colourmap name to use for the bands
         cmap_limits (tuple): fraction of cmap to use (DEFAULT: (0.2, 0.8)).
-        n_colours (int): number of colours to use from cmap (DEFAULT: 4).
         unstacked_pdos (bool): whether to plot projected DOS as stack or overlapping.
         spin_only (str): either 'up' or 'down' to only plot one spin channel.
         preserve_kspace_distance (bool): whether to preserve distances in reciprocal space when
@@ -93,8 +96,9 @@ def plot_spectral(seeds, **kwargs):
                      'phonons': False, 'gap': False,
                      'colour_by_seed': False, 'external_efermi': None,
                      'labels': None, 'cmap': None, 'cmap_limits': (0.2, 0.8), 'band_colour': 'occ',
-                     'n_colours': 4, 'spin_only': None, 'figsize': None, 'filename': None,
+                     'spin_only': None, 'figsize': None, 'filename': None,
                      'pdis_interpolation_factor': 2, 'pdis_point_scale': 25, 'projectors_to_plot': None,
+                     'projector_colours': None,
                      'unstacked_pdos': False, 'preserve_kspace_distance': False,
                      'band_reorder': False, 'title': None, 'show': True,
                      'verbosity': 0, 'highlight_bands': None, 'pdos_hide_sum': True}
@@ -142,11 +146,15 @@ def plot_spectral(seeds, **kwargs):
 
     if kwargs.get('plot_window') is not None:
         if isinstance(kwargs.get('plot_window'), list):
-            if len(kwargs.get('plot_window')) != 2:
-                exit('plot_window must have length 2 or be a single number')
-            kwargs['plot_window'] = sorted(kwargs.get('plot_window'))
+            if len(kwargs.get('plot_window')) == 1:
+                kwargs['plot_window'] = (-kwargs['plot_window'][0], kwargs['plot_window'][0])
+            elif len(kwargs.get("plot_window")) != 2:
+                raise RuntimeError(f"`plot_window` must have length 2 or be a single number, not {kwargs['plot_window']}")
         else:
-            kwargs['plot_window'] = (-kwargs.get('plot_window'), kwargs.get('plot_window'))
+            kwargs['plot_window'] = (-kwargs['plot_window'], kwargs['plot_window'])
+
+        kwargs['plot_window'] = sorted(kwargs.get('plot_window'))
+
     else:
         kwargs['plot_window'] = None
 
@@ -162,14 +170,19 @@ def plot_spectral(seeds, **kwargs):
             kwargs['plot_pdos'] = any([any([os.path.isfile('{}.{}'.format(seed, ext)) for ext in exts]) for seed in seeds])
 
     figsize = kwargs['figsize']
+    # need to rasterize pdis plots, which we'll force to have negative zorders
+    rasterized = kwargs.get('plot_pdis', False)
+    subplot_kw = {"rasterized": rasterized}
+    if rasterized:
+        subplot_kw["rasterization_zorder"] = -1e3
     if kwargs['plot_bandstructure'] and not kwargs['plot_dos']:
         if figsize is None:
             figsize = (7, 6)
-        fig, ax_dispersion = plt.subplots(figsize=figsize)
[…]
