### 2020-08-06 7203cd8
DOCS: update rtd notebook and changelog
