### 2020-08-06 8fe6bf1
FIX: rasterize at scatter not figure level, due to issues when exporting to PDF
