### 2020-08-06 f1c0a9b
DOCS: changelog formatting tweaks
