### 2020-08-07 cdbca51
MAINT: factored common workflow steps into new submodule
