+++ matador/compute/compute.py
@@ -791,14 +791,17 @@ class ComputeTask:
             AssertionError: if required key is missing.
 
         """
-        for keyword in forbidden:
-            if keyword in calc_doc:
-                del calc_doc[keyword]
+        if forbidden is not None:
+            for keyword in forbidden:
+                if keyword in calc_doc:
+                    del calc_doc[keyword]
 
         failures = []
-        for keyword in required:
-            if keyword not in calc_doc:
-                failures.append(keyword)
+        if required is not None:
+            for keyword in required:
+                if keyword not in calc_doc:
+                    failures.append(keyword)
+
         if failures:
             raise InputError('The following keywords are required for workflow: {}'.format(', '.join(failures)))
 
