+++ matador/workflows/castep/elastic.py
@@ -13,16 +13,17 @@ import logging
 import numpy as np
 from matador.workflows import Workflow
 from matador.crystal.elastic import get_equation_of_state
+from matador.workflows.castep.commom import castep_prerelax
 
 LOG = logging.getLogger('run3')
 
 
-def castep_elastic(relaxer, calc_doc, seed, **kwargs):
+def castep_elastic(computer, calc_doc, seed, **kwargs):
     """ Perform a calculation of the elastic tensor on a system.
     Currently only calculation of the bulk modulus is implemented.
 
     Parameters:
-        relaxer (:obj:`ComputeTask`): the object that will be calling CASTEP.
+        computer (:obj:`ComputeTask`): the object that will be calling CASTEP.
         calc_doc (dict): dictionary of structure and calculation
             parameters.
         seed (str): root seed for the calculation.
@@ -34,7 +35,7 @@ def castep_elastic(relaxer, calc_doc, seed, **kwargs):
         bool: True if Workflow completed successfully, or False otherwise.
 
     """
-    workflow = CastepElasticWorkflow(relaxer, calc_doc, seed, **kwargs)
+    workflow = CastepElasticWorkflow(computer, calc_doc, seed, **kwargs)
     return workflow.success
 
 
@@ -45,7 +46,7 @@ class CastepElasticWorkflow(Workflow):
     of DOS.
 
     Attributes:
-        relaxer (:obj:`ComputeTask`): the object that calls CASTEP.
+        computer (:obj:`ComputeTask`): the object that calls CASTEP.
         calc_doc (dict): the interim dictionary of structural and
             calculation parameters.
         seed (str): the root seed for the calculation.
@@ -86,11 +87,11 @@ class CastepElasticWorkflow(Workflow):
                 f.writelines(results['summary'])
 
 
-def castep_rescaled_volume_scf(relaxer, calc_doc, seed, rescale=1):
+def castep_rescaled_volume_scf(computer, calc_doc, seed, rescale=1):
     """ Run a singleshot SCF calculation.
 
     Parameters:
-        relaxer (:obj:`ComputeTask`): the object that will be calling CASTEP.
+        computer (:obj:`ComputeTask`): the object that will be calling CASTEP.
         calc_doc (dict): the structure to run on.
         seed (str): root filename of structure.
 
@@ -106,29 +107,19 @@ def castep_rescaled_volume_scf(relaxer, calc_doc, seed, rescale=1):
             scf_doc['lattice_cart'][i][k] *= rescale
     scf_doc['task'] = 'singlepoint'
     bulk_mod_seed = seed + '.bulk_mod'
-    relaxer.seed = bulk_mod_seed
+    computer.seed = bulk_mod_seed
 
-    return relaxer.run_castep_singleshot(scf_doc, bulk_mod_seed, keep=True, intermediate=True)
+    return computer.run_castep_singleshot(scf_doc, bulk_mod_seed, keep=True, intermediate=True)
 
 
-def castep_elastic_prerelax(relaxer, calc_doc, seed):
+def castep_elastic_prerelax(computer, calc_doc, seed):
     """ Run a geometry optimisation before re-scaling volumes SCF-style calculation.
 
     Parameters:
-        relaxer (:obj:`ComputeTask`): the object that will be calling CASTEP.
+        computer (:obj:`ComputeTask`): the object that will be calling CASTEP.
         calc_doc (dict): the structure to run on.
         seed (str): root filename of structure.
 
     """
     LOG.info('Performing CASTEP elastic pre-relax...')
-    relax_doc = copy.deepcopy(calc_doc)
-    if 'geom_max_iter' not in relax_doc:
[…]

[…]
