### 2020-08-07 29432a0
ENH: added magres workflow with pre-relax
