### 2020-08-07 b0ef42a
MAINT: more refactoring of workflows
