+++ matador/workflows/castep/__init__.py
@@ -7,7 +7,7 @@ CASTEP-specific implementation, for example electronic spectropscopy or phonons.
 """
 
 
-__all__ = ['castep_full_phonon', 'castep_full_spectral', 'castep_elastic']
+__all__ = ('castep_full_phonon', 'castep_full_spectral', 'castep_elastic', 'castep_full_magres')
 
 __author__ = 'Matthew Evans'
 __maintainer__ = 'Matthew Evans'
@@ -16,3 +16,4 @@ __maintainer__ = 'Matthew Evans'
 from .phonons import castep_full_phonon
 from .spectral import castep_full_spectral
 from .elastic import castep_elastic
+from .magres import castep_full_magres
