### 2020-08-07 1e0f277
ENH: add magres workflow into run3
