+++ tests/test_workflows.py
@@ -12,7 +12,7 @@ import numpy as np
 
 from .utils import MatadorUnitTest, REAL_PATH, detect_program
 from matador.compute import ComputeTask
-from matador.scrapers import cell2dict, param2dict, phonon2dict
+from matador.scrapers import cell2dict, param2dict, phonon2dict, magres2dict
 
 
 HOSTNAME = os.uname()[1]
@@ -32,7 +32,7 @@ else:
 
 @unittest.skipIf(not CASTEP_PRESENT, "CASTEP not found.")
 class ElasticWorkflowTest(MatadorUnitTest):
-    """ Run a spectral workflow calculation. """
+    """ Run a elastic workflow calculation. """
 
     def test_bulk_mod(self):
         for _f in glob.glob(REAL_PATH + "data/elastic_workflow/*"):
@@ -71,7 +71,7 @@ class ElasticWorkflowTest(MatadorUnitTest):
 
 @unittest.skipIf(not CASTEP_PRESENT, "CASTEP not found.")
 class PhononWorkflowTest(MatadorUnitTest):
-    """ Run a spectral workflow calculation. """
+    """ Run a phonon workflow calculation. """
 
     def test_phonon(self):
         for _f in glob.glob(REAL_PATH + "data/phonon_workflow/*"):
@@ -99,8 +99,69 @@ class PhononWorkflowTest(MatadorUnitTest):
         self.assertTrue(os.path.isfile("completed/Si2.phonon_dos"))
 
         phon, s = phonon2dict("completed/Si2.phonon")
+        a = 2.7355124
+        np.testing.assert_array_almost_equal(
+            phon["lattice_cart"],
+            np.array([[0, a, a], [a, 0, a], [a, a, 0]]),
+            decimal=3,
+
+        )
+
+        a = 3.869
+        np.testing.assert_array_almost_equal(
+            phon["lattice_abc"],
+            np.array([[a, a, a], [60, 60, 60]]),
+            decimal=3
+        )
         self.assertTrue(s, msg="Failed to read phonon file")
         self.assertGreater(np.min(phon["eigenvalues_q"]), -0.05)
+
+        self.assertTrue(os.path.isfile("completed/Si2.cell"))
+        self.assertTrue(os.path.isfile("completed/Si2.res"))
+
+
+@unittest.skipIf(not CASTEP_PRESENT, "CASTEP not found.")
+class MagresWorkflowTest(MatadorUnitTest):
+    """ Run a magres workflow calculation. """
+
+    def test_magres(self):
+        for _f in glob.glob(REAL_PATH + "data/magres_workflow/*"):
+            shutil.copy(_f, ".")
+
+        cell_dict, _ = cell2dict("Si.cell", db=False)
+        param_dict, _ = param2dict("Si.param", db=False)
+        _ = ComputeTask(
+            res="Si2.res",
+            ncores=NCORES,
+            nnodes=None,
+            node=None,
+            cell_dict=cell_dict,
+            param_dict=param_dict,
+            verbosity=VERBOSITY,
+            compute_dir="tmpier_tst",
+        )
+
+        self.assertTrue(os.path.isfile("completed/Si2.check"))
+
+        self.assertTrue(os.path.isfile("completed/Si2.bands"))
+        self.assertTrue(os.path.isfile("completed/Si2.castep"))
+        self.assertTrue(os.path.isfile("completed/Si2.magres"))
+
+        magres, s = magres2dict("completed/Si2.magres")
+        self.assertTrue(s, msg="Failed to read magres file")
+
+        a = 3.866895
+        np.testing.assert_array_almost_equal(
+            magres["lattice_abc"],
+            np.array([[a, a, a], [60, 60, 60]]),
+            decimal=3
[…]
