### 2020-08-09 d9a0256
Changelog formatting tweaks (#95)
* DOCS: changelog formatting tweaks

* MAINT: added dos_plot(...) and dispersion_plot(...) to __all__
