### 2020-08-10 1806050
MAINT: refactor bulk modulus plots
