### 2020-08-10 08c63cf
TST: add bulk modulus value tests
