### 2020-08-10 e0356b5
MAINT: make sure bulk modulus workflow uses correct cell even if continuation is off
