### 2020-08-10 7929974
ENH: added elec_energy_tol parameter and enhanced tests for magres workflow
