### 2020-08-10 67e3d0f
MAINT: refactored all scrapers to be less file extension dependent
