+++ matador/scrapers/castep_scrapers.py
@@ -10,22 +10,35 @@ inputs and outputs.
 from collections import defaultdict
 import os
 import glob
-import gzip
 import warnings
 import pwd
 import numpy as np
-from matador.utils.cell_utils import abc2cart, calc_mp_spacing, cart2volume, wrap_frac_coords, cart2abc, cart2frac
+from pathlib import Path
+from matador.utils.cell_utils import (
+    abc2cart,
+    calc_mp_spacing,
+    cart2volume,
+    wrap_frac_coords,
+    cart2abc,
+    cart2frac,
+)
 from matador.utils.chem_utils import get_stoich, INVERSE_CM_TO_EV
-from matador.scrapers.utils import DFTError, ComputationError, scraper_function, f90_float_parse
+from matador.scrapers.utils import (
+    DFTError,
+    ComputationError,
+    scraper_function,
+    f90_float_parse,
+    get_flines_extension_agnostic,
+)
 
 
 @scraper_function
-def res2dict(seed, db=True, **kwargs):
+def res2dict(fname, db=True, **kwargs):
     """ Extract available information from .res file; preferably
     used in conjunction with cell or param file.
 
     Parameters:
-        seed (str or list): filename or list of filenames of res file(s)
+        fname (str or list): filename or list of filenames of res file(s)
             (with or without extension).
 
     Keyword arguments:
@@ -36,22 +49,18 @@ def res2dict(seed, db=True, **kwargs):
             if the scrape was successful.
 
     """
+    flines, fname = get_flines_extension_agnostic(fname, "res")
     res = dict()
-    # read .res file into array
-    if seed.endswith('.res'):
-        seed = seed.replace('.res', '')
-    with open(seed + '.res', 'r') as f:
-        flines = f.readlines()
+
     # add .res to source
-    res['source'] = []
-    res['source'].append(seed + '.res')
+    res['source'] = [fname]
     # grab file owner username
     try:
-        res['user'] = pwd.getpwuid(os.stat(seed + '.res').st_uid).pw_name
+        res['user'] = pwd.getpwuid(os.stat(fname).st_uid).pw_name
     except Exception:
         pass
 
-    get_seed_metadata(res, seed)
+    get_seed_metadata(res, fname)
     # alias special lines in res file
     titl = ''
     cell = ''
@@ -129,12 +138,12 @@ def res2dict(seed, db=True, **kwargs):
 
 
 @scraper_function
-def cell2dict(seed, db=False, lattice=True, positions=True, **kwargs):
+def cell2dict(fname, db=False, lattice=True, positions=True, **kwargs):
     """ Extract available information from .cell file; probably
     to be merged with another dict from a .param or .res file.
 
     Parameters:
-        seed (str/list): filename or list of filenames of cell file(s)
+        fname (str/list): filename or list of filenames of cell file(s)
             to scrape, with or without extension.
 
     Keyword arguments:
@@ -148,13 +157,11 @@ def cell2dict(seed, db=False, lattice=True, positions=True, **kwargs):
 
     """
     cell = dict()
-    if seed.endswith('.cell'):
-        seed = seed.replace('.cell', '')
-    with open(seed + '.cell', 'r') as f:
-        flines = f.readlines()
+    flines, fname = get_flines_extension_agnostic(fname, "cell")
+
     # add cell file to source
-    cell['source'] = []
-    cell['source'].append(seed + '.cell')
+    cell['source'] = [fname]
+
     for line_no, line in enumerate(flines):
         if line.startswith(('#', '!')):
             continue
@@ -439,7 +446,7 @@ def cell2dict(seed, db=False, lattice=True, positions=True, **kwargs):
     if db:
         for species in cell['species_pot']:
[…]
