### 2020-08-10 ef92b9c
ENH: enable EFIELD calculation in phonon workflow if requested
