+++ matador/utils/castep_params.py
@@ -2,7 +2,7 @@
 automatically generated with file_utils.scrape_castep_params().
 """
 
-CASTEP_VERSION = 18.1
+CASTEP_VERSION = 20.1
 
 CASTEP_CELL_KEYWORDS = [
     "atomic_init",
@@ -101,7 +101,10 @@ CASTEP_CELL_BLOCKS = [
     "bs_kpoint_list",
     "bs_kpoint_path",
     "cell_constraints",
+    "cell_devel_code",
     "elnes_kpoint_list",
+    "ep_kpoint_pairs",
+    "external_bfield",
     "external_efield",
     "external_pressure",
     "hubbard_u",
@@ -127,7 +130,9 @@ CASTEP_PARAMS = [
     "backup_interval",
     "basis_de_dloge",
     "basis_precision",
+    "bfield_unit",
     "born_charge_sum_rule",
+    "boundary_type",
     "bs_eigenvalue_tol",
     "bs_max_cg_steps",
     "bs_max_iter",
@@ -139,9 +144,11 @@ CASTEP_PARAMS = [
     "bs_xc_definition",
     "bs_xc_functional",
     "calculate_born_charges",
+    "calculate_deltascf",
     "calculate_densdiff",
     "calculate_elf",
     "calculate_hirshfeld",
+    "calculate_modos",
     "calculate_raman",
     "calculate_stress",
     "calc_full_ex_pot",
@@ -155,12 +162,28 @@ CASTEP_PARAMS = [
     "continuation",
     "cut_off_energy",
     "data_distribution",
+    "deltascf_checkpoint",
+    "deltascf_constraints",
+    "deltascf_dftu_checkpoint",
+    "deltascf_excite",
+    "deltascf_method",
+    "deltascf_mode",
+    "deltascf_net_spin_1",
+    "deltascf_net_spin_2",
+    "deltascf_overlap_cutoff",
+    "deltascf_smearing",
     "devel_code",
+    "dielec_emb_bulk_permittivity",
+    "dielec_emb_cavity_type",
+    "dielec_emb_func_method",
+    "dielec_emb_param_set",
     "dipole_correction",
     "dipole_dir",
     "dipole_unit",
     "efermi_tol",
+    "efield_calculate_nonlinear",
     "efield_calc_ion_permittivity",
+    "efield_chi2_unit",
     "efield_convergence_win",
     "efield_dfpt_method",
     "efield_energy_tol",
@@ -207,6 +230,8 @@ CASTEP_PARAMS = [
     "ga_mutate_amp",
     "ga_mutate_rate",
     "ga_pop_size",
+    "ga_spin_mutate_amp",
+    "ga_spin_mutate_rate",
     "geom_convergence_win",
     "geom_disp_tol",
     "geom_energy_tol",
@@ -217,12 +242,25 @@ CASTEP_PARAMS = [
     "geom_max_iter",
     "geom_method",
     "geom_modulus_est",
+    "geom_preconditioner",
+    "geom_precon_exp_a",
+    "geom_precon_exp_c_stab",
+    "geom_precon_exp_mu",
+    "geom_precon_exp_r_cut",
+    "geom_precon_exp_r_nn",
+    "geom_precon_ff_c_stab",
+    "geom_precon_ff_r_cut",
+    "geom_precon_scale_cell",
     "geom_spin_fix",
     "geom_stress_tol",
     "geom_tpsd_init_stepsize",
     "geom_tpsd_iterchange",
     "geom_use_linmin",
     "grid_scale",
+    "implicit_solvent_apolar_factor",
+    "implicit_solvent_apolar_term",
+    "implicit_solvent_rescale_apolar",
+    "implicit_solvent_surface_tension",
[…]
