### 2020-08-10 0e54f41
DOCS: updated changelog
