### 2020-08-10 baae3de
Added workflow for NMR calculations and refactored older workflows (#96)
Summary:

- New workflow for NMR calculations and refactoring of old workflows (#96).
    - New workflow performs relaxation and high-quality SCF before NMR calculation.
    - Old workflows refactored and improved to enforce certain required parameters for e.g. checkpointing.
    - Enabled phonon workflow for CASTEP ``PHONON+EFIELD`` task.
    - Made file scrapers less dependent on file type.

Commits:

* MAINT: factored common workflow steps into new submodule

* ENH: added magres workflow with pre-relax

* TST: add test
[…]
