+++ tests/test_workflows.py
@@ -12,7 +12,7 @@ import numpy as np
 
 from .utils import MatadorUnitTest, REAL_PATH, detect_program
 from matador.compute import ComputeTask
-from matador.scrapers import cell2dict, param2dict, phonon2dict
+from matador.scrapers import cell2dict, param2dict, phonon2dict, magres2dict
 
 
 HOSTNAME = os.uname()[1]
@@ -32,7 +32,7 @@ else:
 
 @unittest.skipIf(not CASTEP_PRESENT, "CASTEP not found.")
 class ElasticWorkflowTest(MatadorUnitTest):
-    """ Run a spectral workflow calculation. """
+    """ Run a elastic workflow calculation. """
 
     def test_bulk_mod(self):
         for _f in glob.glob(REAL_PATH + "data/elastic_workflow/*"):
@@ -49,16 +49,29 @@ class ElasticWorkflowTest(MatadorUnitTest):
             param_dict=param_dict,
             verbosity=VERBOSITY,
             compute_dir="/tmp/scratch_test",
-            workflow_kwargs={"plot": False, "num_volumes": 3},
+            workflow_kwargs={"plot": False, "num_volumes": 5},
         )
 
         self.assertFalse(os.path.isfile("completed/Si2.bib"))
-        self.assertFalse(os.path.isfile("completed/Si2.check"))
+        self.assertTrue(os.path.isfile("completed/Si2.check"))
 
         self.assertTrue(os.path.isfile("completed/Si2.bulk_mod.results"))
         self.assertTrue(os.path.isfile("completed/Si2.bulk_mod.res"))
         self.assertTrue(os.path.isfile("completed/Si2.bulk_mod.castep"))
-        self.assertFalse(os.path.isfile("completed/Si2.bulk_mod.png"))
+
+        with open("completed/Si2.bulk_mod.results", "r") as f:
+            flines = f.readlines()
+
+        B = []
+        for line in flines:
+            if "bulk modulus" in line:
+                B.append(float(line.split()[3]))
+
+        # check all computed bulk mods are between 88-92
+        self.assertEqual(len(B), 3)
+        self.assertTrue(all(abs(b - 90) < 2) for b in B)
+
+        self.assertFalse(os.path.isfile("completed/Si2.bulk_mod.pdf"))
 
         self.assertTrue(os.path.isfile("completed/Si2.res"))
         self.assertTrue(os.path.isfile("completed/Si2.geom"))
@@ -71,7 +84,7 @@ class ElasticWorkflowTest(MatadorUnitTest):
 
 @unittest.skipIf(not CASTEP_PRESENT, "CASTEP not found.")
 class PhononWorkflowTest(MatadorUnitTest):
-    """ Run a spectral workflow calculation. """
+    """ Run a phonon workflow calculation. """
 
     def test_phonon(self):
         for _f in glob.glob(REAL_PATH + "data/phonon_workflow/*"):
@@ -99,8 +112,75 @@ class PhononWorkflowTest(MatadorUnitTest):
         self.assertTrue(os.path.isfile("completed/Si2.phonon_dos"))
 
         phon, s = phonon2dict("completed/Si2.phonon")
+        a = 2.7355124
+        np.testing.assert_array_almost_equal(
+            phon["lattice_cart"], np.array([[0, a, a], [a, 0, a], [a, a, 0]]), decimal=3
+        )
+
+        a = 3.869
+        np.testing.assert_array_almost_equal(
+            phon["lattice_abc"], np.array([[a, a, a], [60, 60, 60]]), decimal=3
+        )
         self.assertTrue(s, msg="Failed to read phonon file")
         self.assertGreater(np.min(phon["eigenvalues_q"]), -0.05)
+
[…]
