### 2020-08-10 20e77df
MAINT: changed default legend position in dos plot when not making bandstructure
