### 2020-08-10 342e2ec
MAINT: changed colour flag to --colours instead of --projector_colours
