### 2020-08-10 af288c8
Updated spectral plot defaults (#100)
* MAINT: changed default legend position in dos plot when not making bandstructure

* MAINT: changed colour flag to --colours instead of --projector_colours

* DOCS: updated changelog
