+++ matador/plotting/spectral_plotting.py
@@ -58,7 +58,6 @@ def plot_spectral(seeds, **kwargs):
         phonons (bool): whether to plot phonon or electronic data
         labels (list): list of strings for legend labels for multiple bandstructures
         gap (bool): whether to draw on the band gap
-        colour_by_seed (bool): plot with a separate colour per bandstructure
         external_efermi (float or list): replace scraped Fermi energy with this value (eV) (can be
             specified per spin channel).
         highlight_bands (list): list of integer indices, colour the bands
@@ -98,7 +97,7 @@ def plot_spectral(seeds, **kwargs):
                      'labels': None, 'cmap': None, 'cmap_limits': (0.2, 0.8), 'band_colour': 'occ',
                      'spin_only': None, 'figsize': None, 'filename': None,
                      'pdis_interpolation_factor': 2, 'pdis_point_scale': 25, 'projectors_to_plot': None,
-                     'projector_colours': None,
+                     'projector_colours': None, 'colours': None,
                      'unstacked_pdos': False, 'preserve_kspace_distance': False,
                      'band_reorder': False, 'title': None, 'show': True,
                      'verbosity': 0, 'highlight_bands': None, 'pdos_hide_sum': True}
@@ -110,22 +109,22 @@ def plot_spectral(seeds, **kwargs):
 
     kwargs['projectors_to_plot'] = _parse_projectors_list(kwargs['projectors_to_plot'])
 
-    if kwargs.get('cmap') is None:
+    if kwargs["projector_colours"] is not None:
+        kwargs['colours'] = kwargs["projector_colours"]
+
+    if kwargs["colours"] is None:
         kwargs['colours'] = list(plt.rcParams['axes.prop_cycle'].by_key()['color'])
+
+    if kwargs.get('cmap') is None:
         plt.rcParams['axes.prop_cycle'] = cycler('color', kwargs['colours'])
     else:
-        if isinstance(kwargs['cmap'], str):
-            print('Adjusting colour palette... to {}'.format(kwargs.get('cmap')))
-            try:
-                kwargs['colours'] = plt.cm.get_cmap(kwargs.get('cmap')).colors
-                plt.rcParams['axes.prop_cycle'] = cycler('color', kwargs['colours'])
-            except AttributeError:
-                kwargs['colours'] = list(plt.rcParams['axes.prop_cycle'].by_key()['color'])
-            kwargs['_mpl_cmap'] = plt.get_cmap(kwargs.get('cmap'))
-        elif isinstance(kwargs['cmap'], list):
-            print('Reading list of colours {}...'.format(kwargs.get('cmap')))
-            kwargs['colours'] = kwargs['cmap']
+        print('Adjusting colour palette... to {}'.format(kwargs.get('cmap')))
+        try:
+            kwargs['colours'] = plt.cm.get_cmap(kwargs.get('cmap')).colors
             plt.rcParams['axes.prop_cycle'] = cycler('color', kwargs['colours'])
+        except AttributeError:
+            kwargs['colours'] = list(plt.rcParams['axes.prop_cycle'].by_key()['color'])
+        kwargs['_mpl_cmap'] = plt.get_cmap(kwargs.get('cmap'))
 
     if (kwargs.get('phonons') and kwargs['band_colour'] == 'occ' and kwargs.get('cmap') is None):
         kwargs['band_colour'] = 'grey'
@@ -139,6 +138,7 @@ def plot_spectral(seeds, **kwargs):
         seeds = [seeds]
 
     if len(seeds) > 1:
+        kwargs["colour_by_seed"] = True
         if kwargs['plot_pdis'] or kwargs['plot_dos']:
             kwargs['plot_pdos'] = False
[…]
