+++ CHANGELOG.rst
@@ -21,6 +21,7 @@ Changes since last release [06/08/2020]
     - Made file scrapers less dependent on file type.
 
 - Updated CASTEP parameter list to 20.1 (#97).
+- Tweaked spectral plotting defaults, including ``--colours`` flag to dispersion script (#98).
 
 
 New in release (0.9.7) [29/07/2020]
