### 2020-08-10 32ee934
Changelog formatting
