### 2020-08-10 ce36569
v0.9.8 (#101)
- Improvements to PDIS functionality (#94).

  - Rasterized scatter points for more efficient exporting and fewer graphical artifacts
  - Made underlying :func:`matador.plotting.spectral_plotting.dos_plot` and :func:`matador.plotting.spectral_plotting.dispersion_plot` more API friendly, and added example notebook.
  - Fixed bug in cell scraping for old ``BS_*`` style keywords.

- Improvements to magres functionality, including scraping of units (#90)
- Example notebooks that do not need external data/databases are now run as part of CI (#91).
- New workflow for NMR calculations a
[…]
