### 2020-08-11 #2563 @whedon
[REVIEW]: matador: a Python library for analysing, curating and performing high-throughput density-functional theory calculations
**Submitting author:** @ml-evs (<a href="http://orcid.org/0000-0002-1182-9098">Matthew L Evans</a>)
**Repository:** <a href="https://github.com/ml-evs/matador" target ="_blank">https://github.com/ml-evs/matador</a>
**Version:** 0.9.9

---
