### 2020-08-11 #2563 @ml-evs
Thanks for initiating this review! I've been implementing a few more features since I submitted, so the version to review should be changed to v0.9.8. Is that a problem?
