### 2020-08-12 bac94b7
ENH: Added optional spectral step to magres workflow
