### 2020-08-12 f4206da
MAINT: moved some spectral workflow specialisations out to the general Workflow
