### 2020-08-21 3f4cbb2
DOCS: set default target for internal referencing
