### 2020-08-21 e705c16
MAINT: make sure Crystals do not end up storing defaultdict instead of dict
