### 2020-08-21 e9363ef
ENH: Added MagresReferencer class to perform linear fits between expt. and theoretical chemical shifts
Set underlying shifts with referencer, and added tests

allow passing axis to MagresReferencer plot

switch to weighted-least-squares and added warning if gradient deviates from -1

added plot+prediction plot

Added .magres as a root_source filetype

Switch to dict for data entry over list

Fix magres plot axis limits and labels

Restructure crystal site data
