+++ matador/magres/referencer.py
@@ -1,0 +1,174 @@
+# coding: utf-8
+# Distributed under the terms of the MIT license.
+
+from typing import Dict, List, Optional
+import numpy as np
+import warnings
+
+from matador.plotting.plotting import plotting_function
+from matador.crystal import Crystal
+
+__all__ = ("MagresReferencer", )
+
+
+class MagresReferencer:
+    """ Class for referencing NMR predictions with experimental data. """
+
+    def __init__(
+        self,
+        structures_exp: Dict[str, Crystal],
+        shifts_exp: List[Dict[str, List[float]]],
+        species: str,
+        structures: Optional[List[Crystal]] = None,
+        warn_tolerance: float = 0.1,
+    ):
+        self.structures_exp = structures_exp
+        self.shifts_exp = shifts_exp
+        self.species = species
+        self.structures = structures
+        self.warn_tolerance = warn_tolerance
+
+        self._calc_shifts = []
+        self._expt_shifts = []
+        self._fit_weights = []
+        self._fit_structures = []
+        self._fitted = False
+
+        for formula in self.shifts_exp:
+            if self.species not in self.shifts_exp[formula]:
+                continue
+
+            if formula not in self.structures_exp:
+                warnings.warn(f"Missing {formula} in reference structures.")
+
+            self.match_exp_structure_shifts(
+                self.structures_exp[formula], self.shifts_exp[formula][self.species]
+            )
+
+        self.fit()
+        self.print_fit_summary()
+
+        if self.structures is not None:
+            self.set_shifts_from_fit(self.structures)
+
+    def match_exp_structure_shifts(self, structure, shifts):
+        relevant_sites = [site for site in structure if site.species == self.species]
+        calc_shifts = sorted(
+            [site["chemical_shielding_iso"] for site in relevant_sites]
+        )
+        _shifts = shifts
+        if (
+            len(_shifts) <= len(relevant_sites)
+            and len(relevant_sites) % len(_shifts) == 0
+        ):
+            multiplier = len(relevant_sites) // len(_shifts)
+            _shifts = [shift for cell in [_shifts] * multiplier for shift in cell]
+            _weights = [1 / multiplier for shift in _shifts]
+        else:
+            raise RuntimeError(
+                f"Incompatible shift sizes: {len(relevant_sites)} (theor.) vs {len(_shifts)} (expt.), "
+                "please provide commensurate cells."
+            )
+
+        _shifts = sorted(_shifts, reverse=True)
+
+        self._calc_shifts.extend(calc_shifts)
+        self._expt_shifts.extend(_shifts)
+        self._fit_weights.extend(_weights)
+        self._fit_structures.extend(len(_shifts) * [structure.formula_tex])
+
+        return _shifts, _weights, calc_shifts
+
+    def set_shifts_from_fit(self, structures):
+        for ind, struc in enumerate(structures):
+            for jnd, site in enumerate(struc):
+                if site.species == self.species:
+                    site["chemical_shift_iso"] = self.predict(site["chemical_shielding_iso"])[0]
+
+    def fit(self):
+        import statsmodels.api as sm
+
+        _fit_shifts = sm.add_constant(self._calc_shifts)
+        self.fit_model = sm.regression.linear_model.WLS(self._expt_shifts, _fit_shifts, weights=self._fit_weights)
+        self.fit_results = self.fit_model.fit()
+        self.fit_intercept = self.fit_results.params[0]
[…]
