+++ matador/magres/__init__.py
@@ -1,0 +1,6 @@
+# coding: utf-8
+# Distributed under the terms of the MIT license.
+
+from .referencer import MagresReferencer
+
+__all__ = ("MagresReferencer",)
