### 2020-08-25 f31cd40
MAINT: construct DOS from PDOS if missing, and correct OptaDOS spin PDOS bug if detected
