### 2020-08-25 5d66833
DOCS: removed defunct old example
