### 2020-08-25 b574d9e
BLD: pinned networkx and refactored requirements
