+++ requirements/optional_requirements.txt
@@ -1,3 +1,4 @@
 pymatgen>=2019.10.4
 ase>=3.18
 ascii_graph>=1.2
+networkx<=2.4
