### 2020-08-25 5ed078e
ENH: added support for CASTEP kpoint path `BREAK` directive (#107)
