### 2020-08-25 9b59f3c
MAINT: refactored spectral plotting and fixed issue with projector colours
