+++ matador/plotting/spectral_plotting.py
@@ -10,6 +10,9 @@ bandstructures for electronic and vibrational calculations.
 import os
 import copy
 import numpy as np
+
+import matplotlib.pyplot as plt
+
 from matador.utils.viz_utils import get_element_colours
 from matador.plotting.plotting import plotting_function
 from matador.scrapers import optados2dict, phonon2dict, bands2dict, phonon_dos2dict
@@ -24,12 +27,10 @@ from matador.utils.chem_utils import INVERSE_CM_TO_EV
 __all__ = ['plot_spectral', 'dos_plot', 'dispersion_plot']
 
 PROJECTOR_MIN = 1e-5
-SPIN_UP_COLOUR = "firebrick"
-SPIN_DOWN_COLOUR = "dodgerblue"
 
 
 @plotting_function
-def plot_spectral(seeds, **kwargs):
+def plot_spectral(seeds, **options):
     """ This function wraps all of the spectral plotting capability of matador.
     When provided with a seed, or seeds, several files will be checked:
 
@@ -79,6 +80,9 @@ def plot_spectral(seeds, **kwargs):
             appear rarefied or compressed in particular regions.
         pdis_interpolation_factor (float): multiple by which to interpolate pDIS bands
         pdis_point_scale (float): size of points in pDIS (DEFAULT: 50).
+        colours (list of str): list of matplotlib colours to override default colour cycle for projectors and otherwise.
+        spin_up_colour (str): matplotlib colour to override the spin-up DOS colour.
+        spin_down_colour (str): matplotlib colour to override the spin-up DOS colour.
         projectors_to_plot (str): comma-separted list of projectors to
             plot in the PDIS or PDOS, provided as element:orbital, e.g.
             "K:s,K:p,P" will plot s and p orbitals for K, and all orbitals for P.
@@ -88,7 +92,6 @@ def plot_spectral(seeds, **kwargs):
             regions where the PDOS is negative (leading to total DOS lower than stacked PDOS) (DEFAULT: False).
 
     """
-    import matplotlib.pyplot as plt
     from cycler import cycler
     # set defaults and update class with desired values
     prop_defaults = {'plot_bandstructure': True, 'plot_dos': True, 'plot_pdos': True, 'plot_pdis': True,
@@ -100,82 +103,86 @@ def plot_spectral(seeds, **kwargs):
                      'projector_colours': None, 'colours': None,
                      'unstacked_pdos': False, 'preserve_kspace_distance': False,
                      'band_reorder': False, 'title': None, 'show': True,
-                     'verbosity': 0, 'highlight_bands': None, 'pdos_hide_sum': True}
+                     'verbosity': 0, 'highlight_bands': None, 'pdos_hide_sum': True,
+                     "spin_up_colour": "firebrick", "spin_down_colour": "dodgerblue"}
 
-    for key in kwargs:
-        if kwargs[key] is not None:
-            prop_defaults[key] = kwargs[key]
-    kwargs = prop_defaults
+    for key in options:
+        if options[key] is not None:
+            prop_defaults[key] = options[key]
+    options = prop_defaults
 
-    kwargs['projectors_to_plot'] = _parse_projectors_list(kwargs['projectors_to_plot'])
+    options['projectors_to_plot'] = _parse_projectors_list(options['projectors_to_plot'])
+    options["colour_cycle"] = tuple(plt.rcParams['axes.prop_cycle'].by_key()['color'])
 
-    if kwargs["projector_colours"] is not None:
-        kwargs['colours'] = kwargs["projector_colours"]
+    if options["projector_colours"] is not None:
+        options["colours"] = options["projector_colours"]
 
-    if kwargs["colours"] is None:
[…]
