### 2020-08-25 c3f6d5b
MAINT: blackened spectral_plotting
