+++ matador/plotting/spectral_plotting.py
@@ -18,13 +18,16 @@ from matador.plotting.plotting import plotting_function
 from matador.scrapers import optados2dict, phonon2dict, bands2dict, phonon_dos2dict
 from matador.scrapers import cell2dict, res2dict
 from matador.orm.spectral import (
-    ElectronicDispersion, ElectronicDOS,
-    VibrationalDispersion, VibrationalDOS,
-    Dispersion, DensityOfStates
+    ElectronicDispersion,
+    ElectronicDOS,
+    VibrationalDispersion,
+    VibrationalDOS,
+    Dispersion,
+    DensityOfStates,
 )
 from matador.utils.chem_utils import INVERSE_CM_TO_EV
 
-__all__ = ['plot_spectral', 'dos_plot', 'dispersion_plot']
+__all__ = ["plot_spectral", "dos_plot", "dispersion_plot"]
 
 PROJECTOR_MIN = 1e-5
 
@@ -93,123 +96,181 @@ def plot_spectral(seeds, **options):
 
     """
     from cycler import cycler
+
     # set defaults and update class with desired values
-    prop_defaults = {'plot_bandstructure': True, 'plot_dos': True, 'plot_pdos': True, 'plot_pdis': True,
-                     'phonons': False, 'gap': False,
-                     'colour_by_seed': False, 'external_efermi': None,
-                     'labels': None, 'cmap': None, 'cmap_limits': (0.2, 0.8), 'band_colour': 'occ',
-                     'spin_only': None, 'figsize': None, 'filename': None,
-                     'pdis_interpolation_factor': 2, 'pdis_point_scale': 25, 'projectors_to_plot': None,
-                     'projector_colours': None, 'colours': None,
-                     'unstacked_pdos': False, 'preserve_kspace_distance': False,
-                     'band_reorder': False, 'title': None, 'show': True,
-                     'verbosity': 0, 'highlight_bands': None, 'pdos_hide_sum': True,
-                     "spin_up_colour": "firebrick", "spin_down_colour": "dodgerblue"}
+    prop_defaults = {
+        "plot_bandstructure": True,
+        "plot_dos": True,
+        "plot_pdos": True,
+        "plot_pdis": True,
+        "phonons": False,
+        "gap": False,
+        "colour_by_seed": False,
+        "external_efermi": None,
+        "labels": None,
+        "cmap": None,
+        "cmap_limits": (0.2, 0.8),
+        "band_colour": "occ",
+        "spin_only": None,
+        "figsize": None,
+        "filename": None,
+        "pdis_interpolation_factor": 2,
+        "pdis_point_scale": 25,
+        "projectors_to_plot": None,
+        "projector_colours": None,
+        "colours": None,
+        "unstacked_pdos": False,
+        "preserve_kspace_distance": False,
+        "band_reorder": False,
+        "title": None,
+        "show": True,
+        "verbosity": 0,
+        "highlight_bands": None,
+        "pdos_hide_sum": True,
+        "spin_up_colour": "firebrick",
+        "spin_down_colour": "dodgerblue",
+    }
 
     for key in options:
         if options[key] is not None:
             prop_defaults[key] = options[key]
     options = prop_defaults
 
-    options['projectors_to_plot'] = _parse_projectors_list(options['projectors_to_plot'])
-    options["colour_cycle"] = tuple(plt.rcParams['axes.prop_cycle'].by_key()['color'])
+    options["projectors_to_plot"] = _parse_projectors_list(
+        options["projectors_to_plot"]
+    )
+    options["colour_cycle"] = tuple(plt.rcParams["axes.prop_cycle"].by_key()["color"])
 
     if options["projector_colours"] is not None:
         options["colours"] = options["projector_colours"]
 
[…]
