### 2020-08-25 1a91573
MAINT: remove band_colour=occ setting and deprecate colour_by_seed option
