### 2020-08-26 b9160bb
MAINT: unify phonon PDOS and electronic PDOS
