+++ tests/test_file_io.py
@@ -1266,9 +1266,11 @@ class ScraperMiscTest(MatadorUnitTest):
         self.assertEqual(dos_data["source"], [phonon_fname])
         self.assertEqual(len(dos_data["dos"]), 10001)
         self.assertEqual(len(dos_data["energies"]), 10001)
-        self.assertEqual(len(dos_data["pdos"]), 2)
-        self.assertEqual(len(dos_data["pdos"][("K", None)]), 10001)
-        self.assertEqual(len(dos_data["pdos"][("P", None)]), 10001)
+        self.assertEqual(len(dos_data["pdos"]["energies"]), 10001)
+        self.assertEqual(len(dos_data["pdos"]["pdos"]), 2)
+        self.assertEqual(len(dos_data["pdos"]["projectors"]), 2)
+        self.assertEqual(len(dos_data["pdos"]["pdos"][("K", None, None)]), 10001)
+        self.assertEqual(len(dos_data["pdos"]["pdos"][("P", None, None)]), 10001)
 
         dos, s = phonon_dos2dict(phonon_fname, as_model=True)
         self.assertTrue(isinstance(dos, VibrationalDOS))
