+++ tests/test_plotting.py
@@ -184,8 +184,10 @@ class SpectralPlotTests(unittest.TestCase):
             "K3P",
             "--png",
             "-ph",
-            "--colours grey green blue",
-            "k",
+            "--colours",
+            "grey",
+            "green",
+            "blue",
             "--figsize",
             "10",
             "10",
