### 2020-08-26 66b227b
MAINT: unify phonon PDOS and electronic PDOS
