### 2020-08-26 af0ccaf
MAINT: improved error handling for provenance scraping (#113)
