### 2020-08-31 0048199
MAINT: tweaks to default 2d hull plot: no colourbar, different ticks and axis limits
