### 2020-08-31 #2563 @ml-evs
> I'm just getting started on the code review. For the record, I want to add the same note as in the [pre-review thread](https://github.com/openjournals/joss-reviews/issues/2529#issuecomment-665162592) that I know @ml-evs in the context of another project. This project is [OPTIMADE](https://github.com/Materials-Consortia/OPTIMADE/), a consortium of research groups and organizations in the computational materials field working on a common REST API standard. I don't believe this to be a conflict for the purposes of this review.

Thanks @mkhorton --- just to confirm we are part of the same consortium but haven't worked together directly.
