### 2020-09-07 18d5f22
ENH: scrape EFG values from magres files and compute C_Q/eta_Q (WIP)
