+++ matador/data/constants.py
@@ -15,6 +15,8 @@ ANGSTROM_CUBED_TO_CENTIMETRE_CUBED = 1e-24
 ELECTRON_CHARGE = physical_constants['elementary charge'][0]
 KELVIN_TO_EV = physical_constants['kelvin-electron volt relationship'][0]
 INVERSE_CM_TO_EV = physical_constants['inverse meter-electron volt relationship'][0] * 100
+EFG_AU_BARN_TO_MHZ = 1
+ELECTRIC_QUADRUPOLE_MOMENTS = {}
 
 __all__ = [
     'FARADAY_CONSTANT_Cpermol',
