+++ tests/test_file_io.py
@@ -681,15 +681,11 @@ class CastepScraperTests(MatadorUnitTest):
             error = True
 
     def test_multiple_exts(self):
-        castep_fname = (
-            REAL_PATH + "data/castep_files/Na-edgecase-CollCode10101"
-        )
+        castep_fname = REAL_PATH + "data/castep_files/Na-edgecase-CollCode10101"
         test_dict, s = castep2dict(castep_fname, db=True)
         self.assertTrue(s, msg="Failed entirely, oh dear!\n{}".format(test_dict))
 
-        castep_fname = (
-            REAL_PATH + "data/castep_files/CuP-thermo-test"
-        )
+        castep_fname = REAL_PATH + "data/castep_files/CuP-thermo-test"
         test_dict, s = castep2dict(castep_fname, db=False)
         self.assertTrue(s, msg="Failed entirely, oh dear!\n{}".format(test_dict))
 
@@ -1654,6 +1650,216 @@ class ScraperMiscTest(MatadorUnitTest):
         self.assertEqual(magres_crystal["calculator_version"], "18.1")
 
         self.assertEqual(magres_crystal["magres_units"]["ms"], "ppm")
+        self.assertEqual(magres_crystal["magres_units"]["lattice"], "Angstrom")
+        self.assertEqual(magres_crystal["magres_units"]["atom"], "Angstrom")
+
+    def test_castep_magres_efg(self):
+        magres_fname = REAL_PATH + "data/magres_files/Al2O3.magres"
+        magres_crystal, s = magres2dict(magres_fname, as_model=False, verbosity=5)
+        self.assertTrue(s)
+        # self.assertEqual(len(magres_crystal["atom_types"]), 40)
+        # np.testing.assert_array_equal(
+        #     magres_crystal.lattice_cart,
+        #     np.array(
+        #         [
+        #             [4.1332870000000002, 0.0000000000000000, 0.0000000000000000],
+        #             [-8.9905292805212659e-4, 6.0637949333506347, 0.0000000000000000],
+        #             [2.0677013018922552, 3.3924745014331725e-1, 12.368724395669441],
+        #         ],
+        #     ),
+        # )
+
+        np.testing.assert_almost_equal(
+            magres_crystal["chemical_shielding_isos"],
+            [
+                161.853,
+                183.025,
+                172.432,
+                172.432,
+                167.275,
+                156.212,
+                156.212,
+                179.976,
+                182.155,
+                182.155,
+                182.817,
+                189.255,
+                189.255,
+                182.817,
+                182.155,
+                182.155,
+                179.976,
+                156.212,
+                156.212,
+                167.275,
+                172.432,
+                172.432,
+                183.025,
+                161.853,
+                543.580,
+                543.580,
+                487.296,
+                539.563,
+                489.973,
+                541.225,
+                540.768,
+                540.768,
+                483.439,
+                483.439,
+                540.768,
+                540.768,
+                541.225,
+                489.973,
+                539.563,
+                487.296,
+            ],
+            decimal=2,
+        )
+        np.testing.assert_almost_equal(
+            magres_crystal["chemical_shift_anisos"],
+            [
+                -17.8527,
+                11.7056,
+                -10.1521,
+                -10.1521,
+                17.2892,
+                -22.7995,
+                -22.7995,
+                8.9808,
+                10.5535,
+                10.5535,
+                7.8228,
+                13.6391,
+                13.6391,
+                7.8228,
+                10.5535,
+                10.5535,
+                8.9808,
+                -22.7995,
+                -22.7995,
+                17.2892,
+                -10.1521,
+                -10.1521,
+                11.7056,
+                -17.8527,
+                -22.4242,
+                -22.4242,
+                -41.4922,
+                -22.2261,
+                46.6031,
+                -18.5085,
+                -19.0290,
[…]
