+++ matador/scrapers/magres_scrapers.py
@@ -15,6 +15,7 @@ import numpy as np
 from matador.utils.cell_utils import cart2abc, cart2frac
 from matador.utils.chem_utils import get_stoich
 from matador.scrapers.utils import scraper_function, get_flines_extension_agnostic
+from matador.data.constants import EFG_AU_BARN_TO_MHZ, ELECTRIC_QUADRUPOLE_MOMENTS
 
 
 @scraper_function
@@ -46,21 +47,22 @@ def magres2dict(fname, **kwargs):
                     raise RuntimeError("Something went wrong in reader loop")
                 if split_line[0] == 'units':
                     magres['magres_units'][split_line[1]] = split_line[2]
-                elif 'lattice' in flines[line_no+i]:
-                    lattice = flines[line_no+i].split()[1:]
+                elif 'lattice' in split_line:
+                    lattice = split_line[1:]
                     for j in range(3):
                         magres['lattice_cart'].append([float(elem) for elem in lattice[j*3:(j+1)*3]])
                     magres['lattice_abc'] = cart2abc(magres['lattice_cart'])
-                elif 'atom' in flines[line_no+i]:
-                    atom = flines[line_no+i].split()
+                elif 'atom' in split_line:
+                    atom = split_line
                     magres['atom_types'].append(atom[1])
                     magres['positions_abs'].append([float(elem) for elem in atom[-3:]])
                 i += 1
             break
 
-    magres['num_atoms'] = len(magres['atom_types'])
-    magres['positions_frac'] = cart2frac(magres['lattice_cart'], magres['positions_abs'])
-    magres['stoichiometry'] = get_stoich(magres['atom_types'])
+    if "atom_types" in magres:
+        magres['num_atoms'] = len(magres['atom_types'])
+        magres['positions_frac'] = cart2frac(magres['lattice_cart'], magres['positions_abs'])
+        magres['stoichiometry'] = get_stoich(magres['atom_types'])
 
     for line_no, line in enumerate(flines):
         line = line.lower().strip()
@@ -75,35 +77,44 @@ def magres2dict(fname, **kwargs):
                     raise RuntimeError("Something went wrong in reader loop")
                 if split_line[0] == 'units':
                     magres['magres_units'][split_line[1]] = split_line[2]
-                elif 'sus' in flines[line_no+i]:
-                    sus = flines[line_no+i].split()[1:]
-                    for j in range(3):
-                        magres['susceptibility_tensor'].append([float(val) for val in sus[3*j:3*(j+1)]])
-                elif 'ms' in flines[line_no+i]:
-                    ms = flines[line_no+i].split()[3:]
-                    magres['magnetic_shielding_tensors'].append([])
-                    for j in range(3):
-                        magres['magnetic_shielding_tensors'][-1].append([float(val) for val in ms[3*j:3*(j+1)]])
-                    magres['chemical_shielding_isos'].append(0)
-                    magres['chemical_shift_anisos'].append(0)
-                    magres['chemical_shift_asymmetries'].append(0)
-                    for j in range(3):
-                        magres['chemical_shielding_isos'][-1] += magres['magnetic_shielding_tensors'][-1][j][j] / 3
+                elif 'sus' in split_line:
+                    magres["susceptibility_tensor"] = np.array([float(val) for val in split_line[1:]]).reshape(3, 3)
+
+                elif 'ms' in split_line:
+                    ms = np.array([float(val) for val in split_line[3:]]).reshape(3, 3)
+                    s_iso = np.trace(ms) / 3
 
[…]
