### 2020-09-07 846b711
Add unit conversions and placeholder values for Al and O
