### 2020-09-07 60096b6
Refactor how site data is loaded, including EFG in Site class
