### 2020-09-07 5ebdb83
Refactor handling of lattice inside Site class
