### 2020-09-07 a1a93fb
Tweaked pretty print for crystals and sites
