### 2020-09-07 3629967
Update plot_convergence script and plots
- Blackened module
- Re-enabled force plots alongside energy
