+++ matador/plotting/convergence_plotting.py
@@ -15,11 +15,20 @@ from matador.plotting.plotting import plotting_function, SAVE_EXTS
 from matador.scrapers.castep_scrapers import castep2dict
 from matador.utils.chem_utils import get_formula_from_stoich
 
-__all__ = ['plot_cutoff_kpt_grid']
+__all__ = ["plot_cutoff_kpt_grid"]
 
 
 @plotting_function
-def plot_cutoff_kpt_grid(data, forces=False, max_energy=50, legend=True, colour_by='formula', **kwargs):
+def plot_cutoff_kpt_grid(
+    data,
+    figsize=None,
+    forces=False,
+    max_energy=50,
+    max_force=None,
+    legend=True,
+    colour_by="formula",
+    **kwargs
+):
     """ Create a composite plot of cutoff energy and kpoint convergence.
 
     Parameters:
@@ -32,65 +41,103 @@ def plot_cutoff_kpt_grid(data, forces=False, max_energy=50, legend=True, colour_
 
     """
     import matplotlib.pyplot as plt
+
+    if figsize is None:
+        if forces:
+            figsize = (8, 7)
+        else:
+            figsize = (8, 4)
+
     if forces:
-        _, ax = plt.subplots(2, 2)
+        _, axes = plt.subplots(2, 2, figsize=figsize, sharex="col", sharey="row")
     else:
-        _, ax = plt.subplots(1, 2)
+        _, axes = plt.subplots(1, 2, figsize=figsize, sharey="row")
+
+    if forces:
+        ax = axes[0][0]
+    else:
+        ax = axes[0]
 
     _, lines, labels = plot_field(
         data,
-        field='formation_energy_per_atom',
-        parameter='cut_off_energy',
-        ax=ax[0],
+        field="formation_energy_per_atom",
+        parameter="cut_off_energy",
+        ax=ax,
         colour_by=colour_by,
         max_y=max_energy,
-        reference='last'
+        label_x=not forces,
+        label_y=True,
+        reference="last",
     )
 
     if len(labels) < 30 and legend:
-        plt.figlegend(lines, labels,
-                      loc='upper center', fontsize=10, ncol=5)
+        plt.figlegend(lines, labels, loc="upper center", fontsize=10, ncol=5)
+
+    if forces:
+        ax = axes[0][1]
+    else:
+        ax = axes[1]
 
     plot_field(
         data,
-        field='formation_energy_per_atom',
-        parameter='kpoints_mp_spacing',
-        ax=ax[1],
+        field="formation_energy_per_atom",
+        parameter="kpoints_mp_spacing",
+        ax=ax,
         colour_by=colour_by,
         max_y=max_energy,
-        reference='last'
+        label_x=not forces,
+        label_y=False,
+        reference="last",
     )
 
     if forces:
         plot_field(
             data,
-            field='forces',
-            parameter='cut_off_energy',
-            ax=ax[2],
-            reference='last'
+            field="forces",
+            parameter="cut_off_energy",
+            ax=axes[1][0],
+            max_y=max_force,
+            label_x=True,
+            label_y=True,
+            reference="last",
         )
         plot_field(
             data,
-            field='forces',
-            parameter='kpoints_mp_spacing',
-            ax=ax[3],
-            reference='last'
+            field="forces",
+            parameter="kpoints_mp_spacing",
+            max_y=max_force,
+            ax=axes[1][1],
+            label_x=True,
+            label_y=False,
+            reference="last",
         )
 
+    plt.tight_layout()
+
     if any(kwargs.get(ext) for ext in SAVE_EXTS):
-        fname = 'conv'
+        fname = "conv"
         for ext in SAVE_EXTS:
             if kwargs.get(ext):
[…]
