### 2020-09-08 #2563 @jgostick
Hi @mkhorton and @srmnitc
How is the review coming along?  I see many 'unchecked' check boxes in the list above...are you having any difficulties with this software?
