### 2020-09-08 ml-evs/matador#119 @ml-evs
Add note about test_requirements.txt in installation/testing instructions
> For @ml-evs, only comment for testing is to add a note about `test_requirements.txt` in the testing instructions.

_Originally posted by @mkhorton in https://github.com/openjournals/joss-reviews/issues/2563#issuecomment-689085417_
