### 2020-09-08 #2563 @ml-evs
> For @ml-evs, only comment for testing is to add a note about `test_requirements.txt` in the testing instructions.

Good spot, thanks. Let me know if there's any trouble with CASTEP and/or the integration tests.
