### 2020-09-13 ml-evs/matador#120 @srmnitc
Minor mistake in install instructions on the doc page
This is part of the review at openjournals/joss-reviews#2563

There is a minor mistake in the installattion instruction [here](https://docs.matador.science/en/latest/install.html). It says `conda ijnstall --yes --file requirements/requirements.txt`, where there is a typo in install.
