### 2020-09-13 ml-evs/matador#121 @srmnitc
Binder notebooks do not work
This is part of the review at openjournals/joss-reviews#2563

There are quite a lot of examples, which is really nice. I was trying to run the examples on mybinder, and I get the feeling that they are not really optimized to run on a binder instance. The binder instance is a good way to get started with matador, and I strongly recommend that these notebooks be made a bit more user friendly. I have summarized the issues below-

- Some of the notebooks need CASTEP to run. While this is completely fine, I think you could mark the notebooks that cannot be run on a binder instance.
- I was able to run `magnes_plotting`, `plot_styles` and `pxid_plotting`, however, they through a lot of warnings, often contradicting the specified example options. For example, the `plot_styles` often seem to fall back to the default options.
- In the notebook `pymatgen_and_ase_interface`, `ase` is not installed on binder, and hence it is not possible to run.
- `phonon_results` also throws an error because the folder does not exist.
- There some folders without notebooks. This is completely fine, however, some guidance to the user would be nice.
- Some notebooks lack an explanation of the example.

I completely understand that it is difficult to maintain numerous examples such as this, but since it provides a good entry point for the potential user. The notebook should be completely executable unless otherwise specified.
