### 2020-09-13 ml-evs/matador#122 @srmnitc
Statement of need in documentation
This is part of the review at openjournals/joss-reviews#2563

The JOSS checklist mentions that there should be statement of need in the documentation. However, I was unable to find this.
