### 2020-09-13 ml-evs/matador#123 @srmnitc
JOSS paper remarks
This is part of the review at openjournals/joss-reviews#2563

Thanks for the quite well-written paper. I only have a few minor things; At the beginning of the paper, there is a line which says "approximate solutions to _these_ quantum mechanical equations", however, which equations are being referred to is not clear. I also think that a clear statement of need is not there in the paper. While the advantages of matador are well described, the question of why exactly matador is needed is probably unanswered. I do not think it is understandable from the rest of the text, so we could leave this one up to the editor.
