### 2020-09-13 #2563 @srmnitc
Hi @ml-evs, I have finished my first round of review. Overall, matador looks quite nice and the documentation looks very good. Thanks! I only have minor points that I have raised on the repo. The major one is just with the jupyter notebooks as a lot of them seem to not run on a binder instance. I would not say this would prevent me from checking the examples section on the checklist, but I think it would be great if you could take a look. Another one is with the statement of need in the paper. I think the advantages of matador are clearly described, but what exactly is the research question that matador answers/or helps to answer? Once again, it is not a major issue, maybe @jgostick could decide if the current state is enough. Thanks for the time and effort in developing matador.
