### 2020-09-14 #2563 @ml-evs
> Hi @ml-evs, I have finished my first round of review. Overall, matador looks quite nice and the documentation looks very good. Thanks! I only have minor points that I have raised on the repo. The major one is just with the jupyter notebooks as a lot of them seem to not run on a binder instance. I would not say this would prevent me from checking the examples section on the checklist, but I think it would be great if you could take a look. Another one is with the statement of need in the paper. I think the advantages of matador are clearly described, but what exactly is the research question that matador answers/or helps to answer? Once again, it is not a major issue, maybe @jgostick could decide if the current state is enough. Thanks for the time and effort in developing matador.

Hi @srmnitc, thanks so much for taking the time to review matador. I'll address your comments in more detail in the issues you have raised (I broadly agree with all of them).

Referring specifically to your final point, in the introduction to the paper we outline the research question that _we_ use the code for, namely job management and analysis for high-throughput crystal structure prediction, but the matador library itself should be generally useful to anyone using CASTEP. I'll have a go at making this clearer in the paper and the documentation and invite you to comment further in the specific issue.
