### 2020-09-15 #2563 @mkhorton
So I actually am having issues compiling CASTEP (which usually never gives me issues) to be able to fully run the integration tests. My apologies for this, I'll try and get this sorted ASAP. The specific error I'm seeing is:

`Error: Type mismatch between actual argument at (1) and actual argument at (2)`

in case anyone in this thread has suggestions.
