### 2020-09-16 #2563 @ml-evs
> So I actually am having issues compiling CASTEP (which usually never gives me issues) to be able to fully run the integration tests. My apologies for this, I'll try and get this sorted ASAP. The specific error I'm seeing is:
>
> `Error: Type mismatch between actual argument at (1) and actual argument at (2)`
>
> in case anyone in this thread has suggestions.

That looks like the classic Fortran compiler error when the code is missing an `implicit none`, but I'd be surprised if that was the case... I assume this is with gfortran (and perhaps a more recent one?) I've certainly compiled CASTEP with gfortran 8 with no problems in the recent past. Perhaps you could post your compiler/arch and the error dump I can try to be more helpful (here or elsewhere)?
