### 2020-09-17 8e5e729
Added further notes on testing and development installations (closes #119)
