### 2020-09-17 e1ed8b7
Added further notes on testing and development installations (closes #119)
