### 2020-09-17 ml-evs/matador#121 @ml-evs
Thanks for this thorough review @srmnitc.

Many of these problems are related, e.g. many of the warnings on the notebooks in your second point relate to missing fonts in the Binder instance.

My suggested fix is to make another folder `/interactive_examples` to point the Binder link to, which will contain symlinks to the notebooks that are actually interactive without extra data.

I'll also go through and add some more clarifying descriptions to the notebooks themselves.
