### 2020-09-18 ml-evs/matador#121 @srmnitc
Thanks @ml-evs! I completely agree that separating them to a folder would be a good solution.
