### 2020-09-23 #2563 @mkhorton
Yes, I apologize. It wasn't very helpful of me to include that error without system details! This is the `gfortran` from `gcc version 10.2.0 (Homebrew GCC 10.2.0)` on macOS 10.15.6 with any necessary dependencies installed via homebrew, e.g. cmake, openmpi.

The integration tests aside, I have otherwise completed my review of matador. Matador is a toolbox for running, storing, analyzing and retrieving data generated from first-principles materials modeling techniques. The code is cleanly separated with the intent of each module well-defined and includes a nice CLI interface, robust integration especially with CASTEP (for launching pre-defined workflows, ingesting ('scraping') contents of directories, running jobs via HPC interfaces and the like and storing and retrieving documents via MongoDB). Plotting functionality is particularly impressive and well structured. The code also includes some basic transformations ("swaps"), phase diagram functionality, and fingerprinting (including a very nice PXRD class). Documentation is good with clean docstrings throughout.

While there are clearly several modules still in development or with placeholder code, the existent code as-is easily passes the threshold for publication and features currently described are functional and useful, and passes all the criteria for a JOSS publication.
