### 2020-09-30 #2563 @ml-evs
Thanks for the constructive review @mkhorton.

I should be able to find the time to address @srmnitc's remaining comments this week, in this PR https://github.com/ml-evs/matador/pull/125.
