### 2020-10-01 5528031
FIX: magres plotting conversion from dict to Crystal was not being triggered
