### 2020-10-02 685f86b
Scrape EFG values from magres files and compute C_Q/eta_Q (#115)
* ENH: scrape EFG values from magres files and compute C_Q/eta_Q (WIP)

* Add unit conversions and placeholder values for Al and O

* Refactor how site data is loaded, including EFG in Site class

* Refactor handling of lattice inside Site class

* Tweaked pretty print for crystals and sites

* ENH: Added quadrupole moments for all nuclei

* FIX: magres plotting conversion from dict to Crystal was not being triggered

* MAINT: Update magres example with new printing
