+++ matador/data/magres/__init__.py
@@ -1,0 +1,6 @@
+# coding: utf-8
+# Distributed under the terms of the MIT License.
+
+from .electric_quadrupole_moments import ELECTRIC_QUADRUPOLE_MOMENTS
+
+__all__ = ("ELECTRIC_QUADRUPOLE_MOMENTS", )

[…]
