### 2020-10-02 8640276
Fixes for spectral models, and added a spectral DOS step to magres workflow (#112)
* ENH: Added optional spectral step to magres workflow

* ENH: added "signal_limits" argument to magres plot and allowed multiple rounds of colour cycle

* MAINT: moved some spectral workflow specialisations out to the general Workflow

* FIX: Make sure default elec_energy_tol gets passed through to workflow
