### 2020-10-05 9456517
DOCS: moved examples around and split them into interactive/non-interactive
