+++ docs/src/examples_index.rst
@@ -5,8 +5,6 @@
 Tutorials and example notebooks
 ===============================
 
-Many of these examples can be run interactively at Binder, though some require external data. |Binder|
-
 Real use-cases
 --------------
 
@@ -15,18 +13,32 @@ Real use-cases
 Example Jupyter notebooks
 -------------------------
 
+Interactive
+~~~~~~~~~~~
+
+These interactive examples can be run under Binder; some features that required system configuration (e.g. fonts when plotting) may not work correctly. |Binder|
+
 .. toctree::
     :maxdepth: 1
 
-    notebooks/hull
-    notebooks/matador_plot_styles
-    notebooks/voltage_from_res_files
-    notebooks/projected_spectral
-    notebooks/projected_spectral_as_subplots
-    notebooks/phonon_results
-    notebooks/pxrd_plotting
-    notebooks/magres_plotting
-    notebooks/pymatgen_and_ase_interface
+    notebooks/interactive/matador_plot_styles
+    notebooks/interactive/voltage_from_res_files
+    notebooks/interactive/pxrd_plotting
+    notebooks/interactive/magres_plotting
+    notebooks/interactive/pymatgen_and_ase_interface
+
+Non-interactive
+~~~~~~~~~~~~~~~
+
+These examples require external data, but can be used as example code.
+
+.. toctree::
+    :maxdepth: 1
+
+    notebooks/non-interactive/hull
+    notebooks/non-interactive/phonon_results
+    notebooks/non-interactive/projected_spectral
+    notebooks/non-interactive/projected_spectral_as_subplots
 
 Tutorials
 ---------
@@ -38,7 +50,7 @@ Tutorials
     tutorials/mongo
 
 .. |Binder| image:: https://mybinder.org/badge_logo.svg
- :target: https://mybinder.org/v2/gh/ml-evs/matador/master?filepath=examples
+ :target: https://mybinder.org/v2/gh/ml-evs/matador/master?filepath=docs/src/notebooks/interactive
 
 .. |CuP Binder| image:: https://mybinder.org/badge_logo.svg
    :target: https://mybinder.org/v2/gh/harpaf13/data.copper-phosphides/master?filepath=CuP_results.ipynb
