### 2020-10-05 860bfa2
DOCS: moved examples around and split them into interactive/non-interactive
