### 2020-10-12 09abeb7
Updated paper (following comments in #123):
- Fixed ambiguous wording around "quantum mechanical equations"
- Separated and reworked "Statement of Need" away from technical summary
