+++ paper/paper.md
@@ -29,11 +29,19 @@ bibliography: paper.bib
 
 # Summary
 
-The properties of materials depend heavily on their atomistic structure; knowledge of the possible stable atomic configurations that define a material is required to understand the performance of many technologically and ecologically relevant devices, such as those used for energy storage [@NaP; @CuP]. First-principles crystal structure prediction (CSP) is the art of finding these stable configurations using only quantum mechanics [@JM]. Density-functional theory (DFT) provides a ubiquitous framework for finding approximate solutions to these quantum mechanical equations computationally; calculations using a modern DFT package are sufficiently robust and accurate that insight into real materials can be readily obtained. The computationally intensive work is performed by well-established, low-level software packages, such as CASTEP [@castep] or Quantum Espresso [@qe], which are able to make use of modern high-performance computers. In order to use these codes easily, reliably and reproducibly, many high-level libraries have been developed to create, curate and manipulate the calculations from these low-level workhorses; `matador` is one such framework.
+The properties of materials depend heavily on their atomistic structure; knowledge of the possible stable atomic configurations that define a material is required to understand the performance of many technologically and ecologically relevant devices, such as those used for energy storage [@NaP; @CuP]. First-principles crystal structure prediction (CSP) is the art of finding these stable configurations using only quantum mechanics [@JM]. Density-functional theory (DFT) is a ubiquitous theoretical framework for finding approximate solutions to quantum mechanics; calculations using a modern DFT package are sufficiently robust and accurate that insight into real materials can be readily obtained. The computationally intensive work is performed by well-established, low-level software packages, such as CASTEP [@castep] or Quantum Espresso [@qe], which are able to make use of modern high-performance computers. In order to use these codes easily, reliably and reproducibly, many high-level libraries have been developed to create, curate and manipulate the calculations from these low-level workhorses; `matador` is one such framework.
 
 # Statement of need
 
-`matador` is a Python 3.6+ library and set of command-line tools for performing and analysing high-throughput DFT calculations using the CASTEP [@castep] and Quantum Espresso [@qe] packages. It promotes the use of local databases to increase the reproducibility and reliability of the computational results, and provides tools to create publication-quality plots of phase diagrams, spectral properties and electrochemistry. It is well-tested and fully-documented at [ReadTheDocs](https://matador-db.readthedocs.io), and comes with several tutorials and examples. The package is available on PyPI under the name [`matador-db`](https://pypi.org/project/matador-db). As with many projects, `matador` is built on top of the scientific Python ecosystem of NumPy [@numpy; @numpybook], SciPy [@scipy] and matplotlib [@matplotlib].
+The purpose of `matador` is four-fold: 
+- to promote the use of local databases and high-throughput workflows to increase the reproducibility of the computational results, 
+- to perform reliable analysis of the stability, structure and properties of materials derived from calculations, 
+- to provide tools to create customisable, publication-quality plots of phase diagrams, spectral properties and electrochemistry,
+- to make the above functionality available to those with limited programming experience.
+
+# `matador`
+
+`matador` is a Python 3.6+ library and set of command-line tools for performing and analysing high-throughput DFT calculations using the CASTEP [@castep] and Quantum Espresso [@qe] packages. It is well-tested and fully-documented at [ReadTheDocs](https://matador-db.readthedocs.io), and comes with several tutorials and examples. The package is available on PyPI under the name [`matador-db`](https://pypi.org/project/matador-db). As with many projects, `matador` is built on top of the scientific Python ecosystem of NumPy [@numpy; @numpybook], SciPy [@scipy] and matplotlib [@matplotlib].
 
[…]
