### 2020-10-12 ml-evs/matador#122 @ml-evs
Hi @srmnitc,

I see the check you mention in the JOSS review checklist; whilst there is no explicit "Statement of Need" section in the documentation, I think the README provides an appropriate summary of the package and it's intended audience as it stands. Looking at a few accepted JOSS repos, I can't see any with an explicit section, which I think is what you're asking for?

I'll close this issue on that note, but please feel free to reopen it if you either think it is not clear in the current README, or if my interpretation of the guidelines is incorrect.
