### 2020-10-13 ml-evs/matador#122 @srmnitc
Yeah, I agree with you here @ml-evs . The core functionality section of the documentation is clear enough.
