+++ examples/README.md
@@ -1,0 +1,7 @@
+# Examples
+
+This folder contains an assortment of examples, some as Jupyter notebooks and
+some as tutorials to be followed according to the text files within. A full
+explanation, with links to the appropriate Binder can be found in the [Examples
+section](https://docs.matador.science/en/stable/examples_index.html) of the
+online documentation.
