### 2020-10-13 be1a740
Added note on tutorials
