### 2020-10-13 c897979
Updated NumPy citation to new Nature paper
