+++ paper/paper.bib
@@ -16,20 +16,24 @@
 }
 
 @article{numpy,
-	title = {The {NumPy} {Array}: {A} {Structure} for {Efficient} {Numerical} {Computation}},
-	volume = {13},
-	issn = {1521-9615},
-	url = {http://ieeexplore.ieee.org/document/5725236/},
-	doi = {10/d8k4p9},
-	number = {2},
-	journal = {Computing in Science \& Engineering},
-	author = {van der Walt, Stéfan and Colbert, S. Chris and Varoquaux, Gaël},
-	month = mar,
-	year = {2011},
-	pmid = {1000224770},
-	note = {arXiv: 1102.1523
-ISBN: 1521-9615 VO - 13},
-	pages = {22--30},
+  author  = {Harris, Charles R. and Millman, K. Jarrod and
+            van der Walt, Stéfan J and Gommers, Ralf and
+            Virtanen, Pauli and Cournapeau, David and
+            Wieser, Eric and Taylor, Julian and Berg, Sebastian and
+            Smith, Nathaniel J. and Kern, Robert and Picus, Matti and
+            Hoyer, Stephan and van Kerkwijk, Marten H. and
+            Brett, Matthew and Haldane, Allan and
+            Fernández del Río, Jaime and Wiebe, Mark and
+            Peterson, Pearu and Gérard-Marchant, Pierre and
+            Sheppard, Kevin and Reddy, Tyler and Weckesser, Warren and
+            Abbasi, Hameer and Gohlke, Christoph and
+            Oliphant, Travis E.},
+  title   = {Array programming with {NumPy}},
+  journal = {Nature},
+  year    = {2020},
+  volume  = {585},
+  pages   = {357–362},
+  doi     = {10.1038/s41586-020-2649-2}
 }
 
 @article{castep,
@@ -196,17 +200,6 @@ ISBN: 1521-9615 VO - 13},
 	note = {Number: 3
 Publisher: Nature Publishing Group},
 	pages = {261--272},
-}
-
-@book{numpybook,
-	address = {Austin, Tex.},
-	title = {Guide to {NumPy}},
-	isbn = {978-1-5173-0007-4},
-	language = {English},
-	publisher = {Continuum Press},
-	author = {Oliphant, Travis E},
-	year = {2015},
-	note = {OCLC: 982090469}
 }
 
 @article{seekpath,
