### 2020-10-13 6e4cc4c
Updated Chem Mater citation
