+++ paper/paper.bib
@@ -174,10 +174,11 @@
 @article{CuP,
 	title = {Novel {Phases} of {Copper} {Phosphides} from {Computational} {Structure} {Searches}},
 	url = {https://pubs.acs.org/doi/10.1021/acs.chemmater.0c02054},
-	urldate = {2020-07-23},
 	journal = {Chemistry of Materials},
 	author = {Harper, Angela F. and Evans, Matthew L. and Morris, Andrew J.},
 	month = june,
+  volume = {32},
+  issue = {15},
 	year = {2020},
   doi = {10.1021/acs.chemmater.0c02054},
 }
