### 2020-10-13 ml-evs/matador#121 @ml-evs
Hi @srmnitc, I think #125 closes this now.

I moved split the examples into interactive/non-interactive in the online documentation, added a higher level summary in the examples folder, and fixed the dependencies that were being installed under myBinder.

Unfortunately I'm struggling to get myBinder to launch the image right now (it builds fine, but runs into miscellaneous launch errors) but I'll try again in the morning.

Once again, please feel free to reopen this if you are not satisfied.
