+++ paper/paper.bib
@@ -44,7 +44,7 @@
 	doi = {10/bdfg3f},
 	number = {5/6},
 	journal = {Zeitschrift für Kristallographie - Crystalline Materials},
-	author = {Clark, Stewart J. and Segall, Matthew D. and Pickard, Chris J and Hasnip, Phil J. and Probert, Matt I J and Refson, Keith and Payne, Mike C.},
+	author = {Clark, Stewart J. and Segall, Matthew D. and Pickard, Chris J. and Hasnip, Phil J. and Probert, Matt I. J. and Refson, Keith and Payne, Mike C.},
 	month = jan,
 	year = {2005},
 	note = {ISBN: 1204090744120},
@@ -60,7 +60,7 @@
 	doi = {10/f5xrnj},
 	number = {5},
 	journal = {Computer Physics Communications},
-	author = {Morris, Andrew J. and Nicholls, Rebecca J and Pickard, C J and Yates, Jonathan R.},
+	author = {Morris, Andrew J. and Nicholls, Rebecca J. and Pickard, Chris J. and Yates, Jonathan R.},
 	year = {2014},
 	note = {Publisher: Elsevier B.V.},
 	keywords = {AJM, DOS, electronic structure, theory},
@@ -106,7 +106,7 @@
 	number = {27},
 	urldate = {2020-06-23},
 	journal = {J. Phys.: Condens. Matter},
-	author = {Larsen, Ask Hjorth and Mortensen, Jens Jørgen and Blomqvist, Jakob and Castelli, Ivano E. and Christensen, Rune and Du{\textbackslash}lak, Marcin and Friis, Jesper and Groves, Michael N. and Hammer, Bjørk and Hargus, Cory and Hermes, Eric D. and Jennings, Paul C. and Jensen, Peter Bjerre and Kermode, James and Kitchin, John R. and Kolsbjerg, Esben Leonhard and Kubal, Joseph and Kaasbjerg, Kristen and Lysgaard, Steen and Maronsson, Jón Bergmann and Maxson, Tristan and Olsen, Thomas and Pastewka, Lars and Peterson, Andrew and Rostgaard, Carsten and Schiøtz, Jakob and Schütt, Ole and Strange, Mikkel and Thygesen, Kristian S. and Vegge, Tejs and Vilhelmsen, Lasse and Walter, Michael and Zeng, Zhenhua and Jacobsen, Karsten W.},
+	author = {Larsen, Ask Hjorth and Mortensen, Jens Jørgen and Blomqvist, Jakob and Castelli, Ivano E. and Christensen, Rune and Dułak, Marcin and Friis, Jesper and Groves, Michael N. and Hammer, Bjørk and Hargus, Cory and Hermes, Eric D. and Jennings, Paul C. and Jensen, Peter Bjerre and Kermode, James and Kitchin, John R. and Kolsbjerg, Esben Leonhard and Kubal, Joseph and Kaasbjerg, Kristen and Lysgaard, Steen and Maronsson, Jón Bergmann and Maxson, Tristan and Olsen, Thomas and Pastewka, Lars and Peterson, Andrew and Rostgaard, Carsten and Schiøtz, Jakob and Schütt, Ole and Strange, Mikkel and Thygesen, Kristian S. and Vegge, Tejs and Vilhelmsen, Lasse and Walter, Michael and Zeng, Zhenhua and Jacobsen, Karsten W.},
 	month = jun,
 	year = {2017},
 	note = {Publisher: IOP Publishing},
@@ -153,7 +153,7 @@
 	url = {https://zenodo.org/record/2628066},
 	urldate = {2020-06-23},
 	publisher = {Zenodo},
[…]
