### 2020-10-14 f8acec0
Switch to postBuild over Dockerfile for binder
