+++ matador/data/__init__.py
@@ -1,8 +1,0 @@
-# coding: utf-8
-# Distributed under the terms of the MIT License.
-
-__all__ = ["GSAS_ATOMIC_SCATTERING_COEFFS", "RASPA_ATOMIC_SCATTERING_COEFFS", "ELECTRIC_QUADRUPOLE_MOMENTS"]
-
-from .atomic_scattering.atomic_scattering_factors_gsas import GSAS_ATOMIC_SCATTERING_COEFFS
-from .atomic_scattering.atomic_scattering_factors_raspa import RASPA_ATOMIC_SCATTERING_COEFFS
-from .magres.electric_quadrupole_moments import ELECTRIC_QUADRUPOLE_MOMENTS
