+++ MANIFEST.in
@@ -1,6 +1,7 @@
 recursive-include matador/config *
 recursive-include docs *
-recursive-include matador/data/atomic_scattering *.dat
+recursive-include matador/data/ *.dat
+recursive-include matador/data/ *.csv
 recursive-include requirements/ *
 include matador/data/nouns
 include matador/data/words
