### 2020-10-14 0637d47
Updated MANIFEST and fixed some imports
