### 2020-10-14 75d361d
Added magres data loader test
