### 2020-10-14 b689683
Tweaked default element_colours path for binder
