### 2020-10-14 ml-evs/matador#121 @ml-evs
> The binder instance still seems to have problems. But everything overall seems to be great!

Yes, you're quite right, I've been trying to fix it in #130. Turns out that the Dockerfile approach is not well-supported. I just need to do a final check of every notebook to make sure they work with the new filepaths, then it should be good to go.
