### 2020-10-14 d947f1b
Reorganized examples once more to avoid having to shuttle data
