### 2020-10-14 c14a517
Reorganized examples once more to avoid having to shuttle data
