+++ examples/tutorials/run3_tutorial/run3_geom.rst
@@ -1,0 +1,227 @@
+.. index:: run3_geom
+
+.. _run3_geom:
+
+Example 1: High-throughput geometry optimisations with CASTEP
+-------------------------------------------------------------
+
+.. _ex1:
+
+
+Example 1.1: Using run3 locally
+~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+
+
+In this example, we will suppose that you want to perform a geometry optimisation on several different polymorphs of TiO\ :sub:`2` from the ICSD. The files for this example can be found in ``examples/run3_tutorial``, `here <https://github.com/ml-evs/matador/blob/develop/examples/run3_tutorial>`_.
+
+Setting up the files
+^^^^^^^^^^^^^^^^^^^^
+
+By default, run3 expects the following files in the job folder:
+
+* one ``.res``/SHELX file `per structure` to optimise
+* one ``$seed.cell`` file that contains the CASTEP CELL keywords common to every structure, e.g. pseudopotentials, kpoint spacing, etc.
+* one ``$seed.param`` file that contains the CASTEP PARAM keywords common to every calculation, e.g. ``cut_off_energy``, ``xc_functional``, ``task``, ``geom_max_iter``, etc.
+* any required pseudopotential files.
+
+.. tip:: If you have a database set up with structures from the OQMD these could be obtained via ``matador query --db oqmd_1.1 -f TiO2 --icsd --res``.
+
+.. tip:: Alternatively, you can turn many file types into ``.res`` using the various ``<format>3shx`` scripts (shx standing for SHELX), e.g. ``cell3shx *.cell``.
+
+The job folder should look something like this::
+
+    $ ls
+    O2Ti-OQMD_112497-CollCode171670.res  O2Ti-OQMD_2575-CollCode9852.res     O2Ti-OQMD_7500-CollCode41493.res
+    O2Ti-OQMD_117323-CollCode182578.res  O2Ti-OQMD_3070-CollCode15328.res    O2Ti-OQMD_84685-CollCode97008.res
+    O2Ti-OQMD_13527-CollCode75179.res    O2Ti-OQMD_31247-CollCode657748.res  O2Ti-OQMD_97161-CollCode154036.res
+    O2Ti-OQMD_19782-CollCode154035.res   O2Ti-OQMD_5979-CollCode31122.res    TiO2.cell
+    O2Ti-OQMD_2475-CollCode9161.res      O2Ti-OQMD_7408-CollCode41056.res    TiO2.param
+
+with ``.param`` file containing::
+
+    $ cat TiO2.param
+    task                 : geometryoptimization
+    xc_functional        : LDA 
+    cut_off_energy       : 300.0 eV
+    geom_force_tol       : 0.1
+    spin_polarized       : false
+    fix_occupancy        : false
+    max_scf_cycles       : 100
+    opt_strategy         : speed
+    page_wvfns           : 0
+    perc_extra_bands     : 40
+    num_dump_cycles      : 0
+    backup_interval      : 0
+    geom_method          : LBFGS
+    geom_max_iter        : 300
+    mix_history_length   : 20
+    finite_basis_corr    : 0
+    fixed_npw            : false
+    write_cell_structure : true
+    write_checkpoint     : none
+    write_bib            : false
+    bs_write_eigenvalues : false
+    calculate_stress     : true
+
+and ``.cell`` file containing::
+
+    $ cat TiO2.cell
+    kpoint_mp_spacing: 0.07
+    
+    %block species_pot
+    QC5
+    %endblock species_pot
+
+    symmetry_generate
+    symmetry_tol: 0.01
+    snap_to_symmetry
+
+.. highlight:: bash
+
+
+Calling run3
+^^^^^^^^^^^^
+
[…]
