+++ examples/run3_elastic/run3_elastic.rst
@@ -1,1 +1,0 @@
-../../docs/src/tutorials/run3_elastic.rst
