+++ examples/README.md
@@ -1,7 +1,12 @@
 # Examples
 
 This folder contains an assortment of examples, some as Jupyter notebooks and
-some as tutorials to be followed according to the text files within. A full
-explanation, with links to the appropriate Binder can be found in the [Examples
-section](https://docs.matador.science/en/stable/examples_index.html) of the
-online documentation.
+some as tutorials to be followed according to the text files within. 
+
+A full explanation, with links to the appropriate Binder can be found in the
+[Examples section](https://docs.matador.science/en/stable/examples_index.html)
+of the online documentation.
+
+Examples in the `./interactive/` folder can be run without any external data,
+either locally or with Binder. Those in the `./non-interactive/` folder require
+external data but provide useful code snippets. 
