+++ examples/run3_phonons/run3_phonon.rst
@@ -1,1 +1,0 @@
-../../docs/src/tutorials/run3_phonon.rst
