+++ examples/bandstructure+dos/run3_spectral.rst
@@ -1,1 +1,0 @@
-../../docs/src/tutorials/run3_spectral.rst

[…]
