### 2020-10-15 #2563 @srmnitc
@jgostick @ml-evs With some issues now fixed, I am able to check all the boxes in the review. With that, I would like to recommend matador for publication. It is a well-written module with very helpful examples and documentation. @jgostick  thanks for inviting me to review and @ml-evs thanks for your efforts in fixing all the issues!
