### 2020-10-15 8424382
Add JOSS paper to main branch (#129)
* DOC: init of JOSS paper

* DOCS: first draft of JOSS-like summary

* DOCS: final draft for JOSS

* Updated paper (following comments in #123):

- Fixed ambiguous wording around "quantum mechanical equations"
- Separated and reworked "Statement of Need" away from technical summary

* Fix typo

* Updated NumPy citation to new Nature paper

* Updated Chem Mater citation

* Finalised citations
