+++ paper/paper.bib
@@ -1,0 +1,273 @@
+
+@article{pymatgen,
+	title = {Python {Materials} {Genomics} (pymatgen): {A} robust, open-source python library for materials analysis},
+	volume = {68},
+	issn = {09270256},
+	shorttitle = {Python {Materials} {Genomics} (pymatgen)},
+	url = {https://linkinghub.elsevier.com/retrieve/pii/S0927025612006295},
+	doi = {10/f4mfm4},
+	language = {en},
+	urldate = {2018-11-13},
+	journal = {Computational Materials Science},
+	author = {Ong, Shyue Ping and Richards, William Davidson and Jain, Anubhav and Hautier, Geoffroy and Kocher, Michael and Cholia, Shreyas and Gunter, Dan and Chevrier, Vincent L. and Persson, Kristin A. and Ceder, Gerbrand},
+	month = feb,
+	year = {2013},
+	pages = {314--319},
+}
+
+@article{numpy,
+  author  = {Harris, Charles R. and Millman, K. Jarrod and
+            van der Walt, Stéfan J and Gommers, Ralf and
+            Virtanen, Pauli and Cournapeau, David and
+            Wieser, Eric and Taylor, Julian and Berg, Sebastian and
+            Smith, Nathaniel J. and Kern, Robert and Picus, Matti and
+            Hoyer, Stephan and van Kerkwijk, Marten H. and
+            Brett, Matthew and Haldane, Allan and
+            Fernández del Río, Jaime and Wiebe, Mark and
+            Peterson, Pearu and Gérard-Marchant, Pierre and
+            Sheppard, Kevin and Reddy, Tyler and Weckesser, Warren and
+            Abbasi, Hameer and Gohlke, Christoph and
+            Oliphant, Travis E.},
+  title   = {Array programming with {NumPy}},
+  journal = {Nature},
+  year    = {2020},
+  volume  = {585},
+  pages   = {357–362},
+  doi     = {10.1038/s41586-020-2649-2}
+}
+
+@article{castep,
+	title = {First principles methods using {CASTEP}},
+	volume = {220},
+	issn = {2196-7105},
+	url = {http://www.degruyter.com/view/j/zkri.2005.220.issue-5-6/zkri.220.5.567.65075/zkri.220.5.567.65075.xml},
+	doi = {10/bdfg3f},
+	number = {5/6},
+	journal = {Zeitschrift für Kristallographie - Crystalline Materials},
+	author = {Clark, Stewart J. and Segall, Matthew D. and Pickard, Chris J. and Hasnip, Phil J. and Probert, Matt I. J. and Refson, Keith and Payne, Mike C.},
+	month = jan,
+	year = {2005},
+	note = {ISBN: 1204090744120},
+	keywords = {theory},
+	pages = {567--570},
+}
+
+@article{optados,
+	title = {{OptaDOS}: {A} tool for obtaining density of states, core-level and optical spectra from electronic structure codes},
+	volume = {185},
+	issn = {00104655},
+	url = {http://dx.doi.org/10.1016/j.cpc.2014.02.013},
+	doi = {10/f5xrnj},
+	number = {5},
+	journal = {Computer Physics Communications},
+	author = {Morris, Andrew J. and Nicholls, Rebecca J. and Pickard, Chris J. and Yates, Jonathan R.},
+	year = {2014},
+	note = {Publisher: Elsevier B.V.},
+	keywords = {AJM, DOS, electronic structure, theory},
+	pages = {1477--1485},
+}
+
+@article{NaP,
[…]
