+++ paper/paper.md
@@ -1,0 +1,69 @@
+---
+title: 'matador: a Python library for analysing, curating and performing high-throughput density-functional theory calculations'
+tags:
+  - Python
+  - density-functional theory
+  - ab initio
+  - crystal structure prediction
+  - materials discovery
+  - databases
+  - castep
+  - quantum espresso
+  - mongodb
+authors:
+  - name: Matthew L. Evans
+    orcid:  0000-0002-1182-9098 
+    affiliation: 1
+  - name: Andrew J. Morris
+    orcid: 0000-0001-7453-5698
+    affiliation: 2
+affiliations:
+ - name: Theory of Condensed Matter Group, Cavendish Laboratory, University of Cambridge, J. J. Thomson Avenue, Cambridge, CB3 0HE, U.K.
+   index: 1
+ - name: School of Metallurgy and Materials, University of Birmingham, Edgbaston, Birmingham, B15 2TT, U.K.
+   index: 2
+
+date: July 2020
+bibliography: paper.bib
+---
+
+# Summary
+
+The properties of materials depend heavily on their atomistic structure; knowledge of the possible stable atomic configurations that define a material is required to understand the performance of many technologically and ecologically relevant devices, such as those used for energy storage [@NaP; @CuP]. First-principles crystal structure prediction (CSP) is the art of finding these stable configurations using only quantum mechanics [@JM]. Density-functional theory (DFT) is a ubiquitous theoretical framework for finding approximate solutions to quantum mechanics; calculations using a modern DFT package are sufficiently robust and accurate that insight into real materials can be readily obtained. The computationally intensive work is performed by well-established, low-level software packages, such as CASTEP [@castep] or Quantum Espresso [@qe], which are able to make use of modern high-performance computers. In order to use these codes easily, reliably and reproducibly, many high-level libraries have been developed to create, curate and manipulate the calculations from these low-level workhorses; `matador` is one such framework.
+
+# Statement of need
+
+The purpose of `matador` is fourfold:
+
+- to promote the use of local databases and high-throughput workflows to increase the reproducibility of the computational results, 
+- to perform reliable analysis of the stability, structure and properties of materials derived from calculations, 
+- to provide tools to create customisable, publication-quality plots of phase diagrams, spectral properties and electrochemistry,
+- to make the above functionality available to those with limited programming experience.
+
+# `matador`
+
+`matador` is a Python 3.6+ library and set of command-line tools for performing and analysing high-throughput DFT calculations using the CASTEP [@castep] and Quantum Espresso [@qe] packages. It is well-tested and fully-documented at [ReadTheDocs](https://matador-db.readthedocs.io), and comes with several tutorials and examples. The package is available on PyPI under the name [`matador-db`](https://pypi.org/project/matador-db). As with many projects, `matador` is built on top of the scientific Python ecosystem of NumPy [@numpy], SciPy [@scipy] and matplotlib [@matplotlib].
+
+`matador` has been developed with high-throughput CSP in mind and has found use in the application of CSP to energy storage materials [@NaP; @CuP]; in this use case, a single compositional phase diagram can consist of tens of thousands of structural relaxation calculations. This package is aimed at users of CASTEP or Quantum Espresso who are comfortable with the command-line, yet maybe lack the Python knowledge required to start from scratch with more sophisticated packages. There are many mature packages that provide overlapping functionality with `matador`, the most widespread of which being the Atomic Simulation Environment (ASE) [@ase] and pymatgen [@pymatgen]. A translation layer to and from the structure representation of both of these packages is provided, such that analysis can be reused and combined. 
+
+![Li-Zn-P ternary phase diagram created with matador, plot generated with matplotlib [@matplotlib] and python-ternary [@ternary].\label{fig:hull}](./LiZnP_hull.pdf)
+
+# Overview of functionality
+
[…]
