### 2020-10-15 d204954
MAINT: minor tweaks to unify volume for binary/ternary electrodes
