### 2020-10-16 f069cb2
Fixes battery plot labels (#117)
* FIX: fixed alignment of voltage plot labels and reran examples

* MAINT: minor tweaks to unify volume for binary/ternary electrodes
