### 2020-10-16 f16e2cf
v0.9.9 (#131)
* Up to 0.9.9
* Update recommended citation
