+++ tasks.py
@@ -8,5 +8,10 @@ from invoke import task
 def publish_to_pypi(c, test=False):
     c.run("rm -rf build dist", warn=True)
     c.run("python setup.py sdist bdist_wheel")
-    c.run("twine upload --verbose dist/*" + " --repository-url https://test.pypi.org/legacy/" if test else "")
+    pypi_cmd = "twine upload --verbose dist/*"
+
+    if test:
+        pypi_cmd += " --repository-url https://test.pypi.org/legacy/"
+
+    c.run(pypi_cmd, pty=True)
     c.run("rm -rf build dist", warn=True)
