### 2020-10-16 #2563 @ml-evs
Hi @jgostick,

ORCIDs in the paper are correct, though I don't see them rendering in the preview.

v0.9.9 has just been released https://github.com/ml-evs/matador/releases/tag/0.9.9

The 0.9.9 release is archived at http://doi.org/10.5281/zenodo.4095078

Author list matches, but the title of the archive is just the package name, which seems to be okay looking at other accepted JOSS papers?
