### 2020-10-23 #2563 @ml-evs
Thanks @jgostick, please keep me in mind if you need reviewers in this area.

Thanks again to @mkhorton and @srmnitc for their time and thoughtful reviews.

Regarding the whedon error, the tidier traceback is:

```
/app/vendor/bundle/ruby/2.4.0/bundler/gems/whedon-d14a699185fb/lib/whedon/bibtex_parser.rb:77:in `doi_citation': undefined method `encode' for nil:NilClass (NoMethodError)
from /app/vendor/bundle/ruby/2.4.0/bundler/gems/whedon-d14a699185fb/lib/whedon/bibtex_parser.rb:64:in `make_citation' from /app/vendor/bundle/ruby/2.4.0/bundler/gems/whedon-d14a699185fb/lib/whedon/bibtex_parser.rb:50:in `block in generate_citations'
from /app/vendor/bundle/ruby/2.4.0/gems/bibtex-ruby-5.1.4/lib/bibtex/bibliography.rb:149:in `each' from /app/vendor/bundle/ruby/2.4.0/gems/bibtex-ruby-5.1.4/lib/bibtex/bibliography.rb:149:in `each'
from /app/vendor/bundle/ruby/2.4.0/bundler/gems/whedon-d14a699185fb/lib/whedon/bibtex_parser.rb:43:in `generate_citations' from /app/vendor/bundle/ruby/2.4.0/bundler/gems/whedon-d14a699185fb/lib/whedon/compilers.rb:246:in `crossref_from_markdown'
from /app/vendor/bundle/ruby/2.4.0/bundler/gems/whedon-d14a699185fb/lib/whedon/compilers.rb:21:in `generate_crossref' from /app/vendor/bundle/ruby/2.4.0/bundler/gems/whedon-d14a699185fb/lib/whedon/processor.rb:100:in `compile'
from /app/vendor/bundle/ruby/2.4.0/bundler/gems/whedon-d14a699185fb/bin/whedon:88:in `compile' from /app/vendor/bundle/ruby/2.4.0/gems/thor-0.20.3/lib/thor/command.rb:27:in `run'
from /app/vendor/bundle/ruby/2.4.0/gems/thor-0.20.3/lib/thor/invocation.rb:126:in `invoke_command' from /app/vendor/bundle/ruby/2.4.0/gems/thor-0.20.3/lib/thor.rb:387:in `dispatch'
from /app/vendor/bundle/ruby/2.4.0/gems/thor-0.20.3/lib/thor/base.rb:466:in `start' from /app/vendor/bundle/ruby/2.4.0/bundler/gems/whedon-d14a699185fb/bin/whedon:131:in `<top (required)>'
from /app/vendor/bundle/ruby/2.4.0/bin/whedon:23:in `load' from /app/vendor/bundle/ruby/2.4.0/bin/whedon:23:in `
```

I can't see weird characters that would be missing an encode method in my bib file, but I can try fixing some dodgy indentation that crept in, though I assume I will not be able to test the retrigger the accept command myself!
