### 2020-10-23 99a3fd9
Standardized JOSS bib indentation and URLs
