+++ paper/paper.bib
@@ -1,273 +1,261 @@
-
 @article{pymatgen,
-	title = {Python {Materials} {Genomics} (pymatgen): {A} robust, open-source python library for materials analysis},
-	volume = {68},
-	issn = {09270256},
-	shorttitle = {Python {Materials} {Genomics} (pymatgen)},
-	url = {https://linkinghub.elsevier.com/retrieve/pii/S0927025612006295},
-	doi = {10/f4mfm4},
-	language = {en},
-	urldate = {2018-11-13},
-	journal = {Computational Materials Science},
-	author = {Ong, Shyue Ping and Richards, William Davidson and Jain, Anubhav and Hautier, Geoffroy and Kocher, Michael and Cholia, Shreyas and Gunter, Dan and Chevrier, Vincent L. and Persson, Kristin A. and Ceder, Gerbrand},
-	month = feb,
-	year = {2013},
-	pages = {314--319},
+    title = {Python {Materials} {Genomics} (pymatgen): {A} robust, open-source python library for materials analysis},
+    volume = {68},
+    issn = {09270256},
+    shorttitle = {Python {Materials} {Genomics} (pymatgen)},
+    url = {https://linkinghub.elsevier.com/retrieve/pii/S0927025612006295},
+    doi = {10/f4mfm4},
+    language = {en},
+    urldate = {2018-11-13},
+    journal = {Computational Materials Science},
+    author = {Ong, Shyue Ping and Richards, William Davidson and Jain, Anubhav and Hautier, Geoffroy and Kocher, Michael and Cholia, Shreyas and Gunter, Dan and Chevrier, Vincent L. and Persson, Kristin A. and Ceder, Gerbrand},
+    month = feb,
+    year = {2013},
+    pages = {314--319},
 }
 
 @article{numpy,
-  author  = {Harris, Charles R. and Millman, K. Jarrod and
-            van der Walt, Stéfan J and Gommers, Ralf and
-            Virtanen, Pauli and Cournapeau, David and
-            Wieser, Eric and Taylor, Julian and Berg, Sebastian and
-            Smith, Nathaniel J. and Kern, Robert and Picus, Matti and
-            Hoyer, Stephan and van Kerkwijk, Marten H. and
-            Brett, Matthew and Haldane, Allan and
-            Fernández del Río, Jaime and Wiebe, Mark and
-            Peterson, Pearu and Gérard-Marchant, Pierre and
-            Sheppard, Kevin and Reddy, Tyler and Weckesser, Warren and
-            Abbasi, Hameer and Gohlke, Christoph and
-            Oliphant, Travis E.},
-  title   = {Array programming with {NumPy}},
-  journal = {Nature},
-  year    = {2020},
-  volume  = {585},
-  pages   = {357–362},
-  doi     = {10.1038/s41586-020-2649-2}
+    author  = {Harris, Charles R. and Millman, K. Jarrod and
+              van der Walt, Stéfan J and Gommers, Ralf and
+              Virtanen, Pauli and Cournapeau, David and
+              Wieser, Eric and Taylor, Julian and Berg, Sebastian and
+              Smith, Nathaniel J. and Kern, Robert and Picus, Matti and
+              Hoyer, Stephan and van Kerkwijk, Marten H. and
+              Brett, Matthew and Haldane, Allan and
+              Fernández del Río, Jaime and Wiebe, Mark and
+              Peterson, Pearu and Gérard-Marchant, Pierre and
[…]
