+++ paper/paper.bib
@@ -4,7 +4,7 @@
     issn = {09270256},
     shorttitle = {Python {Materials} {Genomics} (pymatgen)},
     url = {https://linkinghub.elsevier.com/retrieve/pii/S0927025612006295},
-    doi = {10/f4mfm4},
+    doi = {10.1016/j.commatsci.2012.10.028},
     language = {en},
     urldate = {2018-11-13},
     journal = {Computational Materials Science},
@@ -40,7 +40,7 @@
     volume = {220},
     issn = {2196-7105},
     url = {http://www.degruyter.com/view/j/zkri.2005.220.issue-5-6/zkri.220.5.567.65075/zkri.220.5.567.65075.xml},
-    doi = {10/bdfg3f},
+    doi = {10.1524/zkri.220.5.567.65075},
     number = {5/6},
     journal = {Zeitschrift für Kristallographie - Crystalline Materials},
     author = {Clark, Stewart J. and Segall, Matthew D. and Pickard, Chris J. and Hasnip, Phil J. and Probert, Matt I. J. and Refson, Keith and Payne, Mike C.},
@@ -55,7 +55,7 @@
     volume = {185},
     issn = {00104655},
     url = {http://dx.doi.org/10.1016/j.cpc.2014.02.013},
-    doi = {10/f5xrnj},
+    doi = {10.1016/j.cpc.2014.02.013},
     number = {5},
     journal = {Computer Physics Communications},
     author = {Morris, Andrew J. and Nicholls, Rebecca J. and Pickard, Chris J. and Yates, Jonathan R.},
@@ -69,7 +69,7 @@
     volume = {140},
     issn = {0002-7863, 1520-5126},
     url = {http://pubs.acs.org/doi/10.1021/jacs.8b04183},
-    doi = {10/gdq6h4},
+    doi = {10.1021/jacs.8b04183},
     language = {en},
     number = {25},
     urldate = {2018-08-02},
@@ -96,7 +96,7 @@
     volume = {29},
     issn = {0953-8984},
     url = {https://doi.org/10/f9wbtg},
-    doi = {10/f9wbtg},
+    doi = {10.1088/1361-648x/aa680e},
     language = {en},
     number = {27},
     urldate = {2020-06-23},
@@ -114,7 +114,7 @@
     issn = {0953-8984},
     shorttitle = {{QUANTUM} {ESPRESSO}},
     url = {https://doi.org/10/d7spb8},
-    doi = {10/d7spb8},
+    doi = {10.1088/0953-8984/21/39/395502},
     language = {en},
     number = {39},
     urldate = {2020-06-23},
@@ -131,7 +131,7 @@
     volume = {9},
     issn = {1558-366X},
     shorttitle = {Matplotlib},
-    doi = {10/drbjhg},
+    doi = {10.1109/MCSE.2007.55},
     number = {3},
     journal = {Computing in Science Engineering},
     author = {Hunter, John D.},
@@ -183,7 +183,7 @@
     issn = {1548-7105},
     shorttitle = {{SciPy} 1.0},
     url = {https://www.nature.com/articles/s41592-019-0686-2},
-    doi = {10/ggj45f},
+    doi = {10.1038/s41592-019-0686-2},
     language = {en},
     number = {3},
     urldate = {2020-06-23},
@@ -199,7 +199,7 @@
     volume = {128},
     issn = {09270256},
     url = {http://dx.doi.org/10/f9jxbz},
-    doi = {10/f9jxbz},
+    doi = {10.1016/j.commatsci.2016.10.015},
     journal = {Computational Materials Science},
[…]
