### 2020-10-26 #2563 @arfon
@ml-evs - could you merge this PR which modifies the DOIs to the more standard format? https://github.com/ml-evs/matador/pull/132
