### 2020-10-26 #2563 @ml-evs
> @ml-evs - could you merge this PR which modifies the DOIs to the more standard format? [ml-evs/matador#132](https://github.com/ml-evs/matador/pull/132)

Hi @arfon, done. Was ShortDOI the problem?
