### 2020-10-27 #2563 @arfon
> Hi @arfon, done. Was ShortDOI the problem?

Yes, our system currently doesn't support these (although it looks like we should!)
