### 2020-10-27 #2563 @whedon
:wave: @openjournals/joss-eics, this paper is ready to be accepted and published.
